refactor(tests): move public-API vault_sync tests from inline mod to tests/vault_sync_*.rs
Migrates 98 tests out of `src/core/vault_sync.rs::mod tests` into 8
per-feature integration test files under `tests/`, leaving 69 white-box
tests inline with a module-level annotation listing the private items
they exercise. Pure mechanical refactor: no production code, visibility,
or test logic changes — only test relocation with adjusted `use` paths.

New files (each ≤ 1500-LOC budget):
- tests/vault_sync_ipc.rs (8 tests, 303 LOC) — IPC socket + peer auth
- tests/vault_sync_runtime.rs (10 tests, 495 LOC) — start_serve_runtime
  startup/restart orchestration
- tests/vault_sync_handshake.rs (13 tests, 487 LOC) — owner-lease and
  session ack/handshake
- tests/vault_sync_restore.rs (24 tests, 1003 LOC) — begin_restore /
  finalize_pending_restore / restore_reset pipeline
- tests/vault_sync_remap.rs (23 tests, 805 LOC) — remap_collection /
  verify_remap_root / tree-fence source-introspection
- tests/vault_sync_watcher.rs (7 tests, 293 LOC) — collection watcher,
  supervisor, plain_sync reconcile
- tests/vault_sync_audit.rs (3 tests, 291 LOC) — full_hash_audit batching
- tests/vault_sync_misc.rs (10 tests, 485 LOC) — list_memory_collections
  views, ensure_unix_platform, embedding queue, write-allowed gates

Shared helpers (tests/common/vault_sync_fixtures.rs, 254 LOC):
- ENV_MUTATION_LOCK + env_mutation_lock, EnvVarGuard, secure_runtime_root,
  open_test_db, open_test_db_file, insert_collection,
  insert_collection_with_id, insert_page_with_raw_import,
  write_restore_file, production_vault_sync_source,
  manifest_json_for_directory, wait_for_collection_update,
  create_startup_recovery_sentinel, sha256_hex — each used by ≥ 2 of
  the new files.

Inline residue (69 tests, ~3350 LOC, full block 8978 LOC down from 12601):
white-box tests touching private items including IpcPeerCredentials,
current_effective_uid, authorize_server_peer, authorize_client_peer,
IpcHandlerGuard, IPC_HANDLER_LIMIT, WatchEvent, WatcherHandle,
CollectionWatcherState, WatchBatchBuffer, RecoveryInProgressGuard,
LeaseGuard, WriterSideSentinelCrashMode, DEFAULT_FULL_HASH_AUDIT_DAYS,
the #[cfg(all(test, unix))]-gated pub fns (check_fs_precondition,
has_write_dedup, set_force_native_watcher_init_failure), and several
pub(crate) / module-private fns (init_process_registries,
start_short_lived_owner_lease[_with_interval][_for_root_path],
live_serve_endpoint_for_root_path, classify_watch_event,
poll_collection_watcher, process_embedding_job_on_connection,
resume_orphaned_embedding_jobs, run_overflow_recovery_pass,
start_collection_watcher, sync_collection_watchers, watch_callback,
mark_watcher_crashed, publish_watcher_health,
clear_collection_watcher_health_for_test, recovery_sentinel_paths,
register_supervisor_handle, relative_markdown_path, restore_in_progress,
materialize_collection_to_path, infer_restore_relative_path,
ensure_restore_target_is_empty, ensure_restore_not_blocked,
integrity_blocked_label, take_tree_fence, finalize_outcome_label,
configured_embedding_concurrency, configured_full_hash_audit_days,
full_hash_audit_ttl_cutoff, scheduled_full_hash_audit_budget,
detach_active_collections_with_empty_root_path, self_write_dedup_ttl,
self_write_should_suppress_at, sweep_expired_self_write_entries_at,
remember_self_write_path_at, writer_side_dedup_key,
writer_side_dedup_contains, writer_side_sentinel_name,
writer_side_tempfile_name, exercise_writer_side_sentinel_crash_core,
the #[cfg(test)] fn startup_recovery_sentinel_count, plus the
include_str!("vault_sync.rs") path-relative source-introspection.

Verification:
  baseline (parent commit 2370144): passed=1357 failed=0
  this commit:                       passed=1357 failed=0
