+++ src/core/vault_sync.rs
@@ -5932,6 +5932,38 @@ pub fn get_page_by_input(
     })
 }
 
+// reason: white-box; remaining inline tests touch private items including
+// `IpcPeerCredentials`, `current_effective_uid`, `authorize_server_peer`,
+// `authorize_client_peer`, `IPC_HANDLER_LIMIT`, `IpcHandlerGuard`,
+// `WatchEvent`, `WatcherHandle`, `CollectionWatcherState`, `WatchBatchBuffer`,
+// `RecoveryInProgressGuard`, `LeaseGuard`, `WriterSideSentinelCrashMode`,
+// `DEFAULT_FULL_HASH_AUDIT_DAYS`, the `#[cfg(all(test, unix))]`-gated `pub fn`
+// helpers (`check_fs_precondition`, `has_write_dedup`,
+// `set_force_native_watcher_init_failure`), and several `pub(crate)` /
+// module-private fns (`init_process_registries`,
+// `start_short_lived_owner_lease[_with_interval]`,
+// `start_short_lived_owner_lease_for_root_path`,
+// `live_serve_endpoint_for_root_path`, `classify_watch_event`,
+// `poll_collection_watcher`, `process_embedding_job_on_connection`,
+// `resume_orphaned_embedding_jobs`, `run_overflow_recovery_pass`,
+// `start_collection_watcher`, `sync_collection_watchers`, `watch_callback`,
+// `mark_watcher_crashed`, `publish_watcher_health`,
+// `clear_collection_watcher_health_for_test`, `recovery_sentinel_paths`,
+// `register_supervisor_handle`, `relative_markdown_path`, `restore_in_progress`,
+// `materialize_collection_to_path`, `infer_restore_relative_path`,
+// `ensure_restore_target_is_empty`, `ensure_restore_not_blocked`,
+// `integrity_blocked_label`, `take_tree_fence`, `finalize_outcome_label`,
+// `configured_embedding_concurrency`, `configured_full_hash_audit_days`,
+// `full_hash_audit_ttl_cutoff`, `scheduled_full_hash_audit_budget`,
+// `detach_active_collections_with_empty_root_path`,
+// `self_write_dedup_ttl`, `self_write_should_suppress_at`,
+// `sweep_expired_self_write_entries_at`, `remember_self_write_path_at`,
+// `writer_side_dedup_key`, `writer_side_dedup_contains`,
+// `writer_side_sentinel_name`, `writer_side_tempfile_name`,
+// `exercise_writer_side_sentinel_crash_core`, the `#[cfg(test)] fn`
+// `startup_recovery_sentinel_count`, plus the `include_str!("vault_sync.rs")`
+// path-relative source-introspection. Public-API tests live in
+// `tests/vault_sync_*.rs`.
 #[cfg(test)]
 mod tests {
     #![allow(clippy::unwrap_used, clippy::expect_used)]
@@ -6092,114 +6124,6 @@ mod tests {
         page_id
     }
 
-    #[cfg(all(unix, target_os = "linux"))]
-    #[test]
-    fn start_serve_runtime_refuses_insecure_ipc_directory_permissions() {
-        let _env_lock = env_mutation_lock().lock().unwrap();
-        let runtime_root = secure_runtime_root();
-        let socket_dir = runtime_root.path().join("quaid");
-        fs::create_dir_all(&socket_dir).unwrap();
-        fs::set_permissions(&socket_dir, fs::Permissions::from_mode(0o755)).unwrap();
-        let _xdg = EnvVarGuard::set("XDG_RUNTIME_DIR", runtime_root.path().to_str().unwrap());
-        let (_dir, db_path, _conn) = open_test_db_file();
-
-        let error = match start_serve_runtime(db_path) {
-            Ok(_) => panic!("expected insecure socket directory to refuse startup"),
-            Err(error) => error,
-        };
-
-        assert!(matches!(error, VaultSyncError::IpcDirectoryInsecure { .. }));
-        assert!(error.to_string().contains("IpcDirectoryInsecureError"));
-    }
-
-    #[cfg(all(unix, target_os = "linux"))]
-    #[test]
-    fn start_serve_runtime_refuses_insecure_xdg_runtime_root_permissions() {
-        let _env_lock = env_mutation_lock().lock().unwrap();
-        let runtime_root = tempfile::TempDir::new().unwrap();
-        fs::set_permissions(runtime_root.path(), fs::Permissions::from_mode(0o755)).unwrap();
-        let _xdg = EnvVarGuard::set("XDG_RUNTIME_DIR", runtime_root.path().to_str().unwrap());
-        let (_dir, db_path, _conn) = open_test_db_file();
-
-        let error = match start_serve_runtime(db_path) {
-            Ok(_) => panic!("expected insecure XDG runtime root to refuse startup"),
-            Err(error) => error,
-        };
-
-        assert!(matches!(error, VaultSyncError::IpcDirectoryInsecure { .. }));
-        assert!(error.to_string().contains("IpcDirectoryInsecureError"));
-        assert!(error.to_string().contains("mode 755 is not 700"));
-    }
-
-    #[cfg(all(unix, target_os = "linux"))]
-    #[test]
-    fn start_serve_runtime_refuses_insecure_fallback_runtime_root_permissions() {
-        let _env_lock = env_mutation_lock().lock().unwrap();
-        let home = tempfile::TempDir::new().unwrap();
-        let runtime_root = home.path().join(".cache").join("quaid");
-        fs::create_dir_all(&runtime_root).unwrap();
-        fs::set_permissions(&runtime_root, fs::Permissions::from_mode(0o755)).unwrap();
-        let _xdg = EnvVarGuard::clear("XDG_RUNTIME_DIR");
-        let _home = EnvVarGuard::set("HOME", home.path().to_str().unwrap());
-        let (_dir, db_path, _conn) = open_test_db_file();
-
-        let error = match start_serve_runtime(db_path) {
-            Ok(_) => panic!("expected insecure fallback runtime root to refuse startup"),
-            Err(error) => error,
-        };
-
-        assert!(matches!(error, VaultSyncError::IpcDirectoryInsecure { .. }));
-        assert!(error.to_string().contains("IpcDirectoryInsecureError"));
-        assert!(error.to_string().contains(runtime_root.to_str().unwrap()));
-    }
-
-    #[cfg(all(unix, target_os = "linux"))]
-    #[test]
-    fn publish_ipc_socket_unlinks_stale_socket_before_bind() {
-        let _env_lock = env_mutation_lock().lock().unwrap();
-        let runtime_root = secure_runtime_root();
-        let socket_dir = runtime_root.path().join("quaid");
-        fs::create_dir_all(&socket_dir).unwrap();
-        fs::set_permissions(&socket_dir, fs::Permissions::from_mode(0o700)).unwrap();
-        let _xdg = EnvVarGuard::set("XDG_RUNTIME_DIR", runtime_root.path().to_str().unwrap());
-        let (_dir, _db_path, conn) = open_test_db_file();
-        let session_id = "stale-session";
-        conn.execute(
-            "INSERT INTO serve_sessions (session_id, pid, host) VALUES (?1, 42, 'host')",
-            [session_id],
-        )
-        .unwrap();
-        let socket_path = socket_dir.join(format!("{session_id}.sock"));
-        let stale_listener = UnixListener::bind(&socket_path).unwrap();
-        drop(stale_listener);
-
-        let published = publish_ipc_socket(&conn, session_id).unwrap();
-
-        assert_eq!(published.path, socket_path);
-        assert!(published.path.exists());
-        cleanup_published_ipc_socket(&conn, session_id, &published.path).unwrap();
-    }
-
-    #[cfg(all(unix, target_os = "linux"))]
-    #[test]
-    fn audit_bound_ipc_socket_rejects_mode_regression() {
-        let _env_lock = env_mutation_lock().lock().unwrap();
-        let runtime_root = secure_runtime_root();
-        let socket_dir = runtime_root.path().join("quaid");
-        fs::create_dir_all(&socket_dir).unwrap();
-        fs::set_permissions(&socket_dir, fs::Permissions::from_mode(0o700)).unwrap();
-        let _xdg = EnvVarGuard::set("XDG_RUNTIME_DIR", runtime_root.path().to_str().unwrap());
-        let socket_path = socket_dir.join("mode-regression.sock");
-        let listener = UnixListener::bind(&socket_path).unwrap();
-        fs::set_permissions(&socket_path, fs::Permissions::from_mode(0o644)).unwrap();
-
-        let error = audit_bound_ipc_socket(&socket_path).unwrap_err();
-
-        assert!(matches!(error, VaultSyncError::IpcSocketPermission { .. }));
-        drop(listener);
-        let _ = fs::remove_file(&socket_path);
-    }
-
     #[cfg(unix)]
     #[test]
     fn authorize_server_peer_rejects_cross_uid_peer() {
@@ -6239,171 +6163,6 @@ mod tests {
     }
 
     #[test]
-    fn serve_ipc_source_publishes_after_audit_and_cleans_up_before_unregister() {
-        let source = fs::read_to_string(
-            Path::new(env!("CARGO_MANIFEST_DIR"))
-                .join("src")
-                .join("core")
-                .join("vault_sync.rs"),
-        )
-        .unwrap();
-        let publish_start = source
-            .find("fn publish_ipc_socket(")
-            .expect("publish helper present");
-        let publish_end = source[publish_start..]
-            .find("pub fn start_serve_runtime(")
-            .map(|offset| publish_start + offset)
-            .expect("publish helper boundary");
-        let publish_source = &source[publish_start..publish_end];
-        let runtime_root_idx = publish_source
-            .find("ensure_secure_ipc_directory(&location.runtime_root")
-            .expect("secure runtime root check");
-        let dir_idx = publish_source
-            .find("ensure_secure_ipc_directory(&location.socket_dir, true)")
-            .expect("secure socket directory check");
-        let stale_idx = publish_source
-            .find("clear_stale_ipc_socket(&socket_path)")
-            .expect("stale socket cleanup");
-        let set_perms_idx = publish_source
-            .find("fs::set_permissions(&socket_path, fs::Permissions::from_mode(0o600))")
-            .expect("explicit socket permissions set after bind");
-        let audit_idx = publish_source
-            .find("audit_bound_ipc_socket(&socket_path)")
-            .expect("bind-time audit");
-        let publish_idx = publish_source
-            .find("UPDATE serve_sessions SET ipc_path = ?1 WHERE session_id = ?2")
-            .expect("ipc path publish");
-        assert!(
-            runtime_root_idx < dir_idx
-                && dir_idx < stale_idx
-                && stale_idx < set_perms_idx
-                && set_perms_idx < audit_idx
-                && audit_idx < publish_idx
-        );
-
-        let runtime_start = source
-            .find("pub fn start_serve_runtime(")
-            .expect("serve runtime present");
-        let runtime_end = source[runtime_start..]
-            .find("#[derive(Debug, Clone)]")
-            .map(|offset| runtime_start + offset)
-            .expect("serve runtime boundary");
-        let runtime_source = &source[runtime_start..runtime_end];
-        let cleanup_idx = runtime_source
-            .find(
-                "cleanup_published_ipc_socket(&conn, &session_id_for_thread, &published_ipc.path)",
-            )
-            .expect("ipc cleanup call");
-        let unregister_idx = runtime_source
-            .find("unregister_session(&conn, &session_id_for_thread)")
-            .expect("session unregister call");
-        assert!(cleanup_idx < unregister_idx);
-    }
-
-    #[test]
-    fn serve_ipc_source_refuses_cross_uid_peer_before_request_dispatch() {
-        let source = fs::read_to_string(
-            Path::new(env!("CARGO_MANIFEST_DIR"))
-                .join("src")
-                .join("core")
-                .join("vault_sync.rs"),
-        )
-        .unwrap();
-        let handler_start = source
-            .find("fn handle_ipc_client(")
-            .expect("ipc handler present");
-        let handler_end = source[handler_start..]
-            .find("fn write_ipc_response(")
-            .map(|offset| handler_start + offset)
-            .expect("ipc handler boundary");
-        let handler_source = &source[handler_start..handler_end];
-        let peer_idx = handler_source
-            .find("let peer = peer_credentials_for_stream(&stream)?;")
-            .expect("peer credential lookup");
-        let auth_idx = handler_source
-            .find("authorize_server_peer(socket_path, &peer)?;")
-            .expect("server peer auth");
-        let parse_idx = handler_source
-            .find("serde_json::from_str::<IpcRequest>(line.trim_end())")
-            .expect("request parse");
-        assert!(peer_idx < auth_idx && auth_idx < parse_idx);
-    }
-
-    #[test]
-    fn drain_embedding_queue_marks_failed_jobs_and_retries_after_backoff() {
-        let conn = open_test_db();
-        let page_id = insert_page_with_raw_import(
-            &conn,
-            1,
-            "notes/retry",
-            &Uuid::now_v7().to_string(),
-            "Retry candidate truth.",
-            b"---\ntitle: Retry\ntype: note\n---\nRetry candidate truth.\n",
-            "notes/retry.md",
-        );
-        crate::core::raw_imports::enqueue_embedding_job(&conn, page_id).unwrap();
-        conn.execute(
-            "UPDATE embedding_models SET vec_table = 'bad-table' WHERE active = 1",
-            [],
-        )
-        .unwrap();
-
-        drain_embedding_queue(&conn).unwrap();
-
-        let failed_row: (String, i64, Option<String>) = conn
-            .query_row(
-                "SELECT job_state, attempt_count, last_error
-                 FROM embedding_jobs
-                 WHERE page_id = ?1",
-                [page_id],
-                |row| Ok((row.get(0)?, row.get(1)?, row.get(2)?)),
-            )
-            .unwrap();
-        assert_eq!(failed_row.0, "failed");
-        assert_eq!(failed_row.1, 1);
-        assert!(failed_row
-            .2
-            .as_deref()
-            .is_some_and(|message| message.contains("unsafe vec table name")));
-
-        drain_embedding_queue(&conn).unwrap();
-        let attempt_count_after_immediate_retry: i64 = conn
-            .query_row(
-                "SELECT attempt_count FROM embedding_jobs WHERE page_id = ?1",
-                [page_id],
-                |row| row.get(0),
-            )
-            .unwrap();
-        assert_eq!(attempt_count_after_immediate_retry, 1);
-
-        conn.execute(
-            "UPDATE embedding_jobs
-             SET started_at = datetime('now', '-2 seconds')
-             WHERE page_id = ?1",
-            [page_id],
-        )
-        .unwrap();
-        conn.execute(
-            "UPDATE embedding_models
-             SET vec_table = 'page_embeddings_vec_384'
-             WHERE active = 1",
-            [],
-        )
-        .unwrap();
-
-        drain_embedding_queue(&conn).unwrap();
-
-        let remaining_jobs: i64 = conn
-            .query_row(
-                "SELECT COUNT(*) FROM embedding_jobs WHERE page_id = ?1",
-                [page_id],
-                |row| row.get(0),
-            )
-            .unwrap();
-        assert_eq!(remaining_jobs, 0);
-    }
-
-    #[test]
     fn configured_embedding_concurrency_uses_positive_env_override_when_present() {
         let _lock = env_mutation_lock().lock().unwrap();
         let _clear = EnvVarGuard::clear("QUAID_EMBEDDING_CONCURRENCY");
@@ -6424,42 +6183,6 @@ mod tests {
 
         let _invalid = EnvVarGuard::set("QUAID_EMBEDDING_CONCURRENCY", "bogus");
         assert_eq!(configured_embedding_concurrency(), fallback);
-    }
-
-    #[test]
-    fn drain_embedding_queue_leaves_five_attempt_jobs_failed_without_reclaiming() {
-        let conn = open_test_db();
-        let page_id = insert_page_with_raw_import(
-            &conn,
-            1,
-            "notes/permanent-failure",
-            &Uuid::now_v7().to_string(),
-            "Permanent failure truth.",
-            b"---\ntitle: Permanent Failure\ntype: note\n---\nPermanent failure truth.\n",
-            "notes/permanent-failure.md",
-        );
-        conn.execute(
-            "INSERT INTO embedding_jobs (page_id, job_state, attempt_count, last_error, started_at)
-             VALUES (?1, 'failed', 5, 'still broken', '2026-04-28T00:00:00Z')",
-            [page_id],
-        )
-        .unwrap();
-
-        drain_embedding_queue(&conn).unwrap();
-
-        let row: (String, i64, Option<String>, Option<String>) = conn
-            .query_row(
-                "SELECT job_state, attempt_count, last_error, started_at
-                 FROM embedding_jobs
-                 WHERE page_id = ?1",
-                [page_id],
-                |row| Ok((row.get(0)?, row.get(1)?, row.get(2)?, row.get(3)?)),
-            )
-            .unwrap();
-        assert_eq!(row.0, "failed");
-        assert_eq!(row.1, 5);
-        assert_eq!(row.2.as_deref(), Some("still broken"));
-        assert_eq!(row.3.as_deref(), Some("2026-04-28T00:00:00Z"));
     }
 
     #[test]
@@ -6558,65 +6281,6 @@ mod tests {
         assert!(row.2.is_none());
 
         unregister_session(&conn, &session_id).unwrap();
-    }
-
-    #[test]
-    fn list_memory_collections_counts_pending_and_running_separately_from_failed_jobs() {
-        let conn = open_test_db();
-        let root = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", root.path());
-        let pending_page_id = insert_page_with_raw_import(
-            &conn,
-            collection_id,
-            "notes/pending",
-            &Uuid::now_v7().to_string(),
-            "Pending truth.",
-            b"---\ntitle: Pending\ntype: note\n---\nPending truth.\n",
-            "notes/pending.md",
-        );
-        let running_page_id = insert_page_with_raw_import(
-            &conn,
-            collection_id,
-            "notes/running",
-            &Uuid::now_v7().to_string(),
-            "Running truth.",
-            b"---\ntitle: Running\ntype: note\n---\nRunning truth.\n",
-            "notes/running.md",
-        );
-        let failed_page_id = insert_page_with_raw_import(
-            &conn,
-            collection_id,
-            "notes/failed",
-            &Uuid::now_v7().to_string(),
-            "Failed truth.",
-            b"---\ntitle: Failed\ntype: note\n---\nFailed truth.\n",
-            "notes/failed.md",
-        );
-        conn.execute(
-            "INSERT INTO embedding_jobs (page_id, job_state) VALUES (?1, 'pending')",
-            [pending_page_id],
-        )
-        .unwrap();
-        conn.execute(
-            "INSERT INTO embedding_jobs (page_id, job_state, started_at)
-             VALUES (?1, 'running', '2026-04-28T00:00:00Z')",
-            [running_page_id],
-        )
-        .unwrap();
-        conn.execute(
-            "INSERT INTO embedding_jobs (page_id, job_state, attempt_count, last_error)
-             VALUES (?1, 'failed', 5, 'boom')",
-            [failed_page_id],
-        )
-        .unwrap();
-
-        let view = list_memory_collections(&conn)
-            .unwrap()
-            .into_iter()
-            .find(|view| view.name == "work")
-            .unwrap();
-        assert_eq!(view.embedding_queue_depth, 2);
-        assert_eq!(view.failing_jobs, 1);
     }
 
     fn write_restore_file(root: &Path, relative_path: &str, bytes: &[u8]) {
@@ -6903,66 +6567,6 @@ mod tests {
     }
 
     #[test]
-    fn list_memory_collections_only_marks_restore_in_progress_after_release_ack() {
-        let conn = open_test_db();
-        let active_id = insert_collection(&conn, "active", Path::new("vault-active"));
-        let restoring_pending_id = insert_collection(
-            &conn,
-            "restoring-pending",
-            Path::new("vault-restoring-pending"),
-        );
-        let restoring_live_id =
-            insert_collection(&conn, "restoring-live", Path::new("vault-restoring-live"));
-        conn.execute(
-            "UPDATE collections
-             SET state = 'active',
-                 restore_command_id = 'restore-active',
-                 watcher_released_at = '2026-04-25T00:00:00Z'
-             WHERE id = ?1",
-            [active_id],
-        )
-        .unwrap();
-        conn.execute(
-            "UPDATE collections
-             SET state = 'restoring',
-                 restore_command_id = 'restore-pending',
-                 watcher_released_at = NULL
-             WHERE id = ?1",
-            [restoring_pending_id],
-        )
-        .unwrap();
-        conn.execute(
-            "UPDATE collections
-             SET state = 'restoring',
-                 restore_command_id = 'restore-live',
-                 watcher_released_at = '2026-04-25T00:00:00Z'
-             WHERE id = ?1",
-            [restoring_live_id],
-        )
-        .unwrap();
-
-        let views = list_memory_collections(&conn).unwrap();
-
-        let active = views.iter().find(|view| view.name == "active").unwrap();
-        assert_eq!(active.state, "active");
-        assert!(!active.restore_in_progress);
-
-        let restoring_pending = views
-            .iter()
-            .find(|view| view.name == "restoring-pending")
-            .unwrap();
-        assert_eq!(restoring_pending.state, "restoring");
-        assert!(!restoring_pending.restore_in_progress);
-
-        let restoring_live = views
-            .iter()
-            .find(|view| view.name == "restoring-live")
-            .unwrap();
-        assert_eq!(restoring_live.state, "restoring");
-        assert!(restoring_live.restore_in_progress);
-    }
-
-    #[test]
     fn collection_recovery_in_progress_guard_sets_and_clears_flag() {
         init_process_registries().unwrap();
         let collection_id = 77;
@@ -6973,211 +6577,6 @@ mod tests {
             assert!(collection_recovery_in_progress(collection_id));
         }
         assert!(!collection_recovery_in_progress(collection_id));
-    }
-
-    #[test]
-    fn load_collection_by_id_round_trips_optional_restore_metadata() {
-        let conn = open_test_db();
-        let collection_id = insert_collection(&conn, "work", Path::new("vault"));
-        conn.execute(
-            "UPDATE collections
-             SET active_lease_session_id = 'serve-1',
-                 restore_command_id = 'restore-1',
-                 restore_lease_session_id = 'cli-lease',
-                 reload_generation = 9,
-                 watcher_released_session_id = 'serve-1',
-                 watcher_released_generation = 8,
-                 watcher_released_at = '2026-04-28T00:00:00Z',
-                 pending_command_heartbeat_at = '2026-04-28T00:01:00Z',
-                 pending_root_path = 'D:\\restored',
-                 pending_restore_manifest = '{\"entries\":[]}',
-                 restore_command_pid = 42,
-                 restore_command_host = 'host-a',
-                 integrity_failed_at = '2026-04-28T00:02:00Z',
-                 pending_manifest_incomplete_at = '2026-04-28T00:03:00Z',
-                 reconcile_halted_at = '2026-04-28T00:04:00Z',
-                 reconcile_halt_reason = 'duplicate_uuid'
-             WHERE id = ?1",
-            [collection_id],
-        )
-        .unwrap();
-
-        let collection = load_collection_by_id(&conn, collection_id).unwrap();
-
-        assert_eq!(
-            collection.active_lease_session_id.as_deref(),
-            Some("serve-1")
-        );
-        assert_eq!(collection.restore_command_id.as_deref(), Some("restore-1"));
-        assert_eq!(
-            collection.restore_lease_session_id.as_deref(),
-            Some("cli-lease")
-        );
-        assert_eq!(collection.reload_generation, 9);
-        assert_eq!(
-            collection.watcher_released_session_id.as_deref(),
-            Some("serve-1")
-        );
-        assert_eq!(collection.watcher_released_generation, Some(8));
-        assert_eq!(
-            collection.pending_root_path.as_deref(),
-            Some("D:\\restored")
-        );
-        assert_eq!(
-            collection.pending_restore_manifest.as_deref(),
-            Some("{\"entries\":[]}")
-        );
-        assert_eq!(collection.restore_command_pid, Some(42));
-        assert_eq!(collection.restore_command_host.as_deref(), Some("host-a"));
-        assert_eq!(
-            collection.reconcile_halt_reason.as_deref(),
-            Some("duplicate_uuid")
-        );
-    }
-
-    #[test]
-    fn load_collection_by_id_rejects_invalid_collection_state() {
-        let conn = open_test_db();
-        let collection_id = insert_collection(&conn, "work", Path::new("vault"));
-        conn.pragma_update(None, "ignore_check_constraints", true)
-            .unwrap();
-        conn.execute(
-            "UPDATE collections SET state = 'bogus' WHERE id = ?1",
-            [collection_id],
-        )
-        .unwrap();
-
-        let error = load_collection_by_id(&conn, collection_id)
-            .unwrap_err()
-            .to_string();
-
-        assert!(error.contains("invalid collection state"));
-        assert!(error.contains("bogus"));
-    }
-
-    #[cfg(not(unix))]
-    #[test]
-    fn ensure_unix_platform_fails_closed_on_windows() {
-        let error = ensure_unix_platform("quaid collection sync")
-            .unwrap_err()
-            .to_string();
-        assert!(error.contains("UnsupportedPlatformError"));
-        assert!(error.contains("quaid collection sync"));
-    }
-
-    #[test]
-    fn build_restore_manifest_for_directory_sorts_paths_and_hashes_contents() {
-        let root = tempfile::TempDir::new().unwrap();
-        write_restore_file(root.path(), "notes/z.md", b"zeta");
-        write_restore_file(root.path(), "notes/a.md", b"alpha");
-
-        let manifest = build_restore_manifest_for_directory(root.path()).unwrap();
-
-        assert_eq!(
-            manifest
-                .entries
-                .iter()
-                .map(|entry| entry.relative_path.as_str())
-                .collect::<Vec<_>>(),
-            vec!["notes/a.md", "notes/z.md"]
-        );
-        assert_eq!(manifest.entries[0].sha256, sha256_hex(b"alpha"));
-        assert_eq!(manifest.entries[0].size_bytes, 5);
-        assert_eq!(manifest.entries[1].sha256, sha256_hex(b"zeta"));
-        assert_eq!(manifest.entries[1].size_bytes, 4);
-    }
-
-    #[test]
-    fn list_memory_collections_reports_integrity_blocked_variants() {
-        let conn = open_test_db();
-        let manifest_tamper = insert_collection(&conn, "manifest-tamper", Path::new("vault-a"));
-        let manifest_retry = insert_collection(&conn, "manifest-retry", Path::new("vault-b"));
-        let duplicate_uuid = insert_collection(&conn, "duplicate-uuid", Path::new("vault-c"));
-        let trivial = insert_collection(&conn, "trivial", Path::new("vault-d"));
-        let unknown = insert_collection(&conn, "unknown", Path::new("vault-e"));
-        conn.execute(
-            "UPDATE collections
-             SET integrity_failed_at = '2026-04-28T00:00:00Z'
-             WHERE id = ?1",
-            [manifest_tamper],
-        )
-        .unwrap();
-        conn.execute(
-            "UPDATE collections
-             SET pending_manifest_incomplete_at = datetime('now', '-7200 seconds')
-             WHERE id = ?1",
-            [manifest_retry],
-        )
-        .unwrap();
-        conn.execute(
-            "UPDATE collections
-             SET reconcile_halted_at = '2026-04-28T00:00:00Z',
-                 reconcile_halt_reason = 'duplicate_uuid'
-             WHERE id = ?1",
-            [duplicate_uuid],
-        )
-        .unwrap();
-        conn.execute(
-            "UPDATE collections
-             SET reconcile_halted_at = '2026-04-28T00:00:00Z',
-                 reconcile_halt_reason = 'unresolvable_trivial_content'
-             WHERE id = ?1",
-            [trivial],
-        )
-        .unwrap();
-        conn.execute(
-            "UPDATE collections
-             SET reconcile_halted_at = '2026-04-28T00:00:00Z',
-                 reconcile_halt_reason = 'mystery'
-             WHERE id = ?1",
-            [unknown],
-        )
-        .unwrap();
-
-        let views = list_memory_collections(&conn).unwrap();
-
-        assert_eq!(
-            views
-                .iter()
-                .find(|view| view.name == "manifest-tamper")
-                .unwrap()
-                .integrity_blocked
-                .as_deref(),
-            Some("manifest_tampering")
-        );
-        assert_eq!(
-            views
-                .iter()
-                .find(|view| view.name == "manifest-retry")
-                .unwrap()
-                .integrity_blocked
-                .as_deref(),
-            Some("manifest_incomplete_escalated")
-        );
-        assert_eq!(
-            views
-                .iter()
-                .find(|view| view.name == "duplicate-uuid")
-                .unwrap()
-                .integrity_blocked
-                .as_deref(),
-            Some("duplicate_uuid")
-        );
-        assert_eq!(
-            views
-                .iter()
-                .find(|view| view.name == "trivial")
-                .unwrap()
-                .integrity_blocked
-                .as_deref(),
-            Some("unresolvable_trivial_content")
-        );
-        assert!(views
-            .iter()
-            .find(|view| view.name == "unknown")
-            .unwrap()
-            .integrity_blocked
-            .is_none());
     }
 
     #[test]
@@ -7435,214 +6834,6 @@ mod tests {
 
         assert!(matches!(error, VaultSyncError::InvariantViolation { .. }));
         assert!(error.to_string().contains("missing active raw_imports"));
-    }
-
-    #[test]
-    fn verify_remap_root_rejects_invalid_quaidignore_in_new_root() {
-        let conn = open_test_db();
-        let old_root = tempfile::TempDir::new().unwrap();
-        let new_root = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", old_root.path());
-        let raw_bytes =
-            b"---\nmemory_id: 11111111-1111-7111-8111-111111111111\n---\nhello world from note a";
-        insert_page_with_raw_import(
-            &conn,
-            collection_id,
-            "notes/a",
-            "11111111-1111-7111-8111-111111111111",
-            "hello world from note a",
-            raw_bytes,
-            "notes/a.md",
-        );
-        fs::create_dir_all(new_root.path().join("notes")).unwrap();
-        fs::write(new_root.path().join("notes").join("a.md"), raw_bytes).unwrap();
-        fs::write(new_root.path().join(".quaidignore"), "[broken\n").unwrap();
-
-        let collection = load_collection_by_id(&conn, collection_id).unwrap();
-        let error = verify_remap_root(&conn, &collection, new_root.path()).unwrap_err();
-
-        assert!(matches!(error, VaultSyncError::InvariantViolation { .. }));
-        assert!(error
-            .to_string()
-            .contains("invalid .quaidignore in remap root"));
-    }
-
-    #[test]
-    fn restore_reset_reports_each_block_reason_before_terminal_reset() {
-        let conn = open_test_db();
-        let root = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", root.path());
-
-        conn.execute(
-            "UPDATE collections
-             SET pending_manifest_incomplete_at = '2026-04-28T00:00:00Z'
-             WHERE id = ?1",
-            [collection_id],
-        )
-        .unwrap();
-        let manifest_error = restore_reset(&conn, "work").unwrap_err().to_string();
-        assert!(manifest_error.contains("RestoreResetBlockedError"));
-        assert!(manifest_error.contains("manifest_incomplete_retryable"));
-
-        conn.execute(
-            "UPDATE collections
-             SET pending_manifest_incomplete_at = NULL,
-                 state = 'restoring',
-                 pending_root_path = 'D:\\restored'
-             WHERE id = ?1",
-            [collection_id],
-        )
-        .unwrap();
-        let pending_error = restore_reset(&conn, "work").unwrap_err().to_string();
-        assert!(pending_error.contains("RestoreResetBlockedError"));
-        assert!(pending_error.contains("pending_finalize"));
-
-        conn.execute(
-            "UPDATE collections
-             SET pending_root_path = NULL,
-                 state = 'restoring',
-                 needs_full_sync = 1
-             WHERE id = ?1",
-            [collection_id],
-        )
-        .unwrap();
-        let progress_error = restore_reset(&conn, "work").unwrap_err().to_string();
-        assert!(progress_error.contains("RestoreResetBlockedError"));
-        assert!(progress_error.contains("restore_in_progress"));
-
-        conn.execute(
-            "UPDATE collections
-             SET state = 'active',
-                 needs_full_sync = 0
-             WHERE id = ?1",
-            [collection_id],
-        )
-        .unwrap();
-        let clean_error = restore_reset(&conn, "work").unwrap_err().to_string();
-        assert!(clean_error.contains("RestoreResetBlockedError"));
-        assert!(clean_error.contains("no_integrity_failure"));
-    }
-
-    #[test]
-    fn reconcile_reset_reports_missing_collection() {
-        assert!(matches!(
-            reconcile_reset(&open_test_db(), "missing"),
-            Err(VaultSyncError::CollectionNotFound { name }) if name == "missing"
-        ));
-    }
-
-    #[test]
-    fn restore_reset_can_return_collection_to_detached_and_finalize_covers_remaining_outcomes() {
-        let conn = open_test_db();
-        let placeholder_id = insert_collection(&conn, "placeholder", Path::new("vault"));
-        conn.execute(
-            "UPDATE collections
-             SET root_path = '',
-                 state = 'restoring',
-                 integrity_failed_at = '2026-04-28T00:00:00Z'
-             WHERE id = ?1",
-            [placeholder_id],
-        )
-        .unwrap();
-        restore_reset(&conn, "placeholder").unwrap();
-        let placeholder = load_collection_by_id(&conn, placeholder_id).unwrap();
-        assert_eq!(placeholder.state, CollectionState::Detached);
-        assert_eq!(placeholder.root_path, "");
-
-        let no_pending_id = insert_collection(&conn, "no-pending", Path::new("vault-no-pending"));
-        let no_pending = finalize_pending_restore(
-            &conn,
-            no_pending_id,
-            FinalizeCaller::ExternalFinalize {
-                session_id: "serve-1".to_owned(),
-            },
-        )
-        .unwrap();
-        assert_eq!(no_pending, FinalizeOutcome::NoPendingWork);
-
-        let orphan_id = insert_collection(&conn, "orphan", Path::new("vault-orphan"));
-        conn.execute(
-            "UPDATE collections
-             SET state = 'restoring',
-                 restore_command_id = 'restore-1',
-                 pending_root_path = NULL
-             WHERE id = ?1",
-            [orphan_id],
-        )
-        .unwrap();
-        let orphan = finalize_pending_restore(
-            &conn,
-            orphan_id,
-            FinalizeCaller::ExternalFinalize {
-                session_id: "serve-1".to_owned(),
-            },
-        )
-        .unwrap();
-        assert_eq!(orphan, FinalizeOutcome::OrphanRecovered);
-
-        let aborted_id = insert_collection(&conn, "aborted", Path::new("vault-aborted"));
-        conn.execute(
-            "UPDATE collections
-             SET state = 'restoring',
-                 pending_root_path = 'D:\\missing-restore-root',
-                 pending_restore_manifest = '{\"entries\":[]}',
-                 restore_command_id = 'restore-1',
-                 pending_command_heartbeat_at = datetime('now', '-120 seconds')
-             WHERE id = ?1",
-            [aborted_id],
-        )
-        .unwrap();
-        let aborted = finalize_pending_restore(
-            &conn,
-            aborted_id,
-            FinalizeCaller::ExternalFinalize {
-                session_id: "serve-1".to_owned(),
-            },
-        )
-        .unwrap();
-        assert_eq!(aborted, FinalizeOutcome::Aborted);
-    }
-
-    #[test]
-    fn finalize_pending_restore_returns_no_pending_work_when_pending_root_path_is_null() {
-        let conn = open_test_db();
-        let collection_id = insert_collection(&conn, "no-pending-null", Path::new("vault-np"));
-        // insert_collection sets state='active' and pending_root_path defaults to NULL
-        let outcome = finalize_pending_restore(
-            &conn,
-            collection_id,
-            FinalizeCaller::ExternalFinalize {
-                session_id: "serve-null-1".to_owned(),
-            },
-        )
-        .unwrap();
-        assert_eq!(outcome, FinalizeOutcome::NoPendingWork);
-    }
-
-    #[test]
-    fn finalize_pending_restore_returns_orphan_recovered_when_restoring_with_null_pending_root() {
-        let conn = open_test_db();
-        let collection_id = insert_collection(&conn, "orphan-null", Path::new("vault-orphan-null"));
-        conn.execute(
-            "UPDATE collections
-             SET state = 'restoring',
-                 restore_command_id = 'restore-orphan-null',
-                 pending_root_path = NULL
-             WHERE id = ?1",
-            [collection_id],
-        )
-        .unwrap();
-        let outcome = finalize_pending_restore(
-            &conn,
-            collection_id,
-            FinalizeCaller::ExternalFinalize {
-                session_id: "serve-null-2".to_owned(),
-            },
-        )
-        .unwrap();
-        assert_eq!(outcome, FinalizeOutcome::OrphanRecovered);
-        let reverted = load_collection_by_id(&conn, collection_id).unwrap();
-        assert_ne!(reverted.state, CollectionState::Restoring);
     }
 
     #[cfg(unix)]
@@ -8221,54 +7412,6 @@ mod tests {
 
     #[cfg(unix)]
     #[test]
-    fn start_serve_runtime_watcher_reconciles_external_edit_after_debounce() {
-        let (_dir, db_path, conn) = open_test_db_file();
-        let root = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", root.path());
-        write_restore_file(
-            root.path(),
-            "notes/a.md",
-            b"---\ntitle: A\ntype: note\n---\nOriginal body.\n",
-        );
-        sync_collection(&conn, "work").unwrap();
-        drop(conn);
-
-        let runtime = start_serve_runtime(db_path.clone()).unwrap();
-
-        write_restore_file(
-            root.path(),
-            "notes/a.md",
-            b"---\ntitle: A\ntype: note\n---\nUpdated by watcher.\n",
-        );
-
-        let compiled_truth = wait_for_collection_update(
-            &db_path,
-            collection_id,
-            Duration::from_secs(8),
-            |verify, collection_id| {
-                verify
-                    .query_row(
-                        "SELECT compiled_truth
-                         FROM pages
-                         WHERE collection_id = ?1 AND slug = 'notes/a'",
-                        [collection_id],
-                        |row| row.get::<_, String>(0),
-                    )
-                    .ok()
-                    .and_then(|compiled_truth| {
-                        compiled_truth
-                            .contains("Updated by watcher.")
-                            .then_some(compiled_truth)
-                    })
-            },
-        );
-        assert!(compiled_truth.contains("Updated by watcher."));
-
-        drop(runtime);
-    }
-
-    #[cfg(unix)]
-    #[test]
     fn start_serve_runtime_watcher_rejects_path_only_dedup_match() {
         let (_dir, db_path, conn) = open_test_db_file();
         let root = tempfile::TempDir::new().unwrap();
@@ -8370,208 +7513,6 @@ mod tests {
         assert!(compiled_truth.contains("Stale dedup must not block ingest."));
 
         drop(runtime);
-    }
-
-    #[test]
-    fn sync_collection_watchers_production_logic_stays_active_only_and_generation_aware() {
-        let source = production_vault_sync_source();
-        let start = source.find("fn sync_collection_watchers(").unwrap();
-        let end = source[start..]
-            .find("fn run_watcher_reconcile(")
-            .map(|offset| start + offset)
-            .unwrap();
-        let snippet = &source[start..end];
-
-        assert!(
-            snippet.contains("detach_active_collections_with_empty_root_path(conn)?;"),
-            "watcher sync must normalize empty root paths before watching: {snippet}"
-        );
-        assert!(
-            snippet.contains("WHERE state = 'active'"),
-            "watcher sync must only enumerate active collections: {snippet}"
-        );
-        assert!(
-            snippet
-                .contains("watchers.retain(|collection_id, _| active.contains_key(collection_id))"),
-            "watcher sync must drop watchers for non-active collections: {snippet}"
-        );
-        assert!(
-            snippet.contains("state.root_path != root_path")
-                && snippet.contains("state.generation != generation"),
-            "watcher sync must replace watchers when root or reload_generation changes: {snippet}"
-        );
-        assert!(
-            snippet.contains("state.generation = generation;"),
-            "replacement watchers must inherit the new reload_generation: {snippet}"
-        );
-    }
-
-    #[test]
-    fn run_overflow_recovery_pass_production_logic_uses_active_lease_and_active_gate() {
-        let source = production_vault_sync_source();
-        let start = source.find("fn run_overflow_recovery_pass(").unwrap();
-        let end = source[start..]
-            .find("pub fn start_serve_runtime(")
-            .map(|offset| start + offset)
-            .unwrap();
-        let snippet = &source[start..end];
-
-        assert!(
-            snippet.contains("WHERE state = 'active' AND needs_full_sync = 1"),
-            "overflow recovery must gate itself to active collections: {snippet}"
-        );
-        assert!(
-            snippet.contains("FullHashReconcileMode::OverflowRecovery"),
-            "overflow recovery must use the repaired mode label: {snippet}"
-        );
-        assert!(
-            snippet.contains("FullHashReconcileAuthorization::ActiveLease"),
-            "overflow recovery must reuse the active lease authorization: {snippet}"
-        );
-        assert!(
-            snippet.contains("overflow_recovery_skipped_lease_mismatch"),
-            "overflow recovery must warn on lease mismatch rather than bypassing ownership: {snippet}"
-        );
-    }
-
-    #[test]
-    fn start_serve_runtime_logs_scheduled_maintenance_failures() {
-        let source = production_vault_sync_source();
-        let start = source.find("pub fn start_serve_runtime(").unwrap();
-        let end = source[start..]
-            .find("pub fn begin_restore(")
-            .map(|offset| start + offset)
-            .unwrap();
-        let snippet = &source[start..end];
-
-        assert!(
-            snippet.contains("WARN: scheduled_full_hash_audit_failed")
-                && snippet.contains("if let Err(error) = run_full_hash_audit_pass(&conn, &session_id_for_thread)")
-                && snippet.contains("WARN: janitor_sweep_failed")
-                && snippet.contains("if let Err(error) = janitor::run_tick(&conn)")
-                && snippet.contains("WARN: raw_import_ttl_sweep_failed")
-                && snippet.contains("if let Err(error) = sweep_raw_import_ttl(&conn)"),
-            "serve loop must log janitor, scheduled audit, and TTL sweep failures instead of discarding them silently"
-        );
-    }
-
-    #[test]
-    fn start_serve_runtime_spawns_extraction_worker() {
-        let source = production_vault_sync_source();
-        let start = source.find("pub fn start_serve_runtime(").unwrap();
-        let end = source[start..]
-            .find("pub fn begin_restore(")
-            .map(|offset| start + offset)
-            .unwrap();
-        let snippet = &source[start..end];
-
-        assert!(
-            snippet.contains("run_extraction_worker"),
-            "serve runtime must spawn the extraction worker so queued jobs are drained: {snippet}"
-        );
-        assert!(
-            snippet.contains("extractor_handle: Some("),
-            "serve runtime must store the extraction worker handle on ServeRuntime for shutdown: {snippet}"
-        );
-    }
-
-    #[test]
-    fn run_extraction_worker_logs_failures_and_honors_stop_signal() {
-        let source = production_vault_sync_source();
-        let start = source
-            .find("fn run_extraction_worker(")
-            .expect("run_extraction_worker function must exist for the worker thread");
-        let end = source[start..]
-            .find("\nfn ")
-            .or_else(|| source[start..].find("\npub fn "))
-            .map(|offset| start + offset)
-            .unwrap_or(source.len());
-        let snippet = &source[start..end];
-
-        assert!(
-            snippet.contains("Worker::new(&conn, LazySlmRunner::new(), ResolvingFactWriter)"),
-            "worker thread must construct the production extractor::Worker: {snippet}"
-        );
-        assert!(
-            snippet.contains("worker.run_once()"),
-            "worker thread must drive the queue via run_once so the stop signal is honored between polls: {snippet}"
-        );
-        assert!(
-            snippet.contains("stop.load(Ordering::SeqCst)"),
-            "worker thread loop must honor the ServeRuntime stop signal: {snippet}"
-        );
-        assert!(
-            snippet.contains("WARN: extraction_worker_run_failed"),
-            "worker thread must log run failures instead of swallowing them: {snippet}"
-        );
-        assert!(
-            snippet.contains("WARN: extraction_worker_init_failed"),
-            "worker thread must log init failures instead of crashing the daemon: {snippet}"
-        );
-    }
-
-    #[test]
-    fn serve_runtime_drop_joins_extraction_worker_handle() {
-        let source = production_vault_sync_source();
-        let start = source.find("impl Drop for ServeRuntime").unwrap();
-        let end = source[start..]
-            .find("\n}\n")
-            .map(|offset| start + offset + 3)
-            .unwrap();
-        let snippet = &source[start..end];
-
-        assert!(
-            snippet.contains("self.extractor_handle.take()"),
-            "ServeRuntime drop must join the extraction worker handle: {snippet}"
-        );
-    }
-
-    #[test]
-    fn start_collection_watcher_production_logic_keeps_native_first_poll_fallback() {
-        let source = production_vault_sync_source();
-        let start = source.find("fn start_collection_watcher(").unwrap();
-        let end = source[start..]
-            .find("fn sync_collection_watchers(")
-            .map(|offset| start + offset)
-            .unwrap();
-        let snippet = &source[start..end];
-
-        assert!(
-            snippet.contains("watcher_native_init_failed"),
-            "native watcher failures must warn before fallback: {snippet}"
-        );
-        assert!(
-            snippet.contains("PollWatcher::new"),
-            "native watcher failures must fall back to poll mode: {snippet}"
-        );
-        assert!(
-            snippet.contains("WatcherMode::Poll"),
-            "fallback path must record poll mode explicitly: {snippet}"
-        );
-    }
-
-    #[test]
-    fn watcher_supervisor_production_logic_tracks_crash_state_and_backoff() {
-        let source = production_vault_sync_source();
-        let start = source.find("fn mark_watcher_crashed(").unwrap();
-        let end = source[start..]
-            .find("fn publish_watcher_health(")
-            .map(|offset| start + offset)
-            .unwrap();
-        let snippet = &source[start..end];
-
-        assert!(
-            snippet.contains("state.mode = WatcherMode::Crashed;"),
-            "watcher crashes must be surfaced as an explicit crashed mode: {snippet}"
-        );
-        assert!(
-            snippet.contains("state.backoff_until = Some(now + backoff);"),
-            "watcher crashes must record restart backoff: {snippet}"
-        );
-        assert!(
-            snippet.contains("watcher_backoff_duration"),
-            "watcher crashes must use exponential backoff helper: {snippet}"
-        );
     }
 
     #[test]
@@ -8854,34 +7795,6 @@ mod tests {
 
     #[cfg(unix)]
     #[test]
-    fn start_serve_runtime_leaves_restoring_needs_full_sync_for_overflow_worker() {
-        let (_dir, db_path, conn) = open_test_db_file();
-        let root = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", root.path());
-        conn.execute(
-            "UPDATE collections
-             SET state = 'restoring',
-                 needs_full_sync = 1,
-                 integrity_failed_at = '2026-04-28T00:00:00Z'
-             WHERE id = ?1",
-            [collection_id],
-        )
-        .unwrap();
-        drop(conn);
-
-        let runtime = start_serve_runtime(db_path.clone()).unwrap();
-        thread::sleep(Duration::from_millis(1200));
-
-        let verify = Connection::open(&db_path).unwrap();
-        let collection = load_collection_by_id(&verify, collection_id).unwrap();
-
-        assert!(collection.needs_full_sync);
-        assert_eq!(collection.state, CollectionState::Restoring);
-        drop(runtime);
-    }
-
-    #[cfg(unix)]
-    #[test]
     fn watch_callback_marks_collection_needs_full_sync_when_channel_is_full() {
         let (_dir, db_path, conn) = open_test_db_file();
         let root = tempfile::TempDir::new().unwrap();
@@ -8949,215 +7862,6 @@ mod tests {
     }
 
     #[test]
-    fn ensure_collection_write_allowed_refuses_on_restoring_or_needs_full_sync() {
-        let conn = open_test_db();
-        let collection_id = insert_collection(&conn, "work", Path::new("vault"));
-        conn.execute(
-            "UPDATE collections SET state = 'restoring', needs_full_sync = 1 WHERE id = ?1",
-            [collection_id],
-        )
-        .unwrap();
-
-        let error = ensure_collection_write_allowed(&conn, collection_id).unwrap_err();
-        assert!(error.to_string().contains("CollectionRestoringError"));
-        assert!(error.to_string().contains("needs_full_sync=true"));
-    }
-
-    #[test]
-    fn ensure_collection_write_allowed_refuses_when_only_needs_full_sync_is_set() {
-        let conn = open_test_db();
-        let collection_id = insert_collection(&conn, "work", Path::new("vault"));
-        conn.execute(
-            "UPDATE collections SET state = 'active', needs_full_sync = 1 WHERE id = ?1",
-            [collection_id],
-        )
-        .unwrap();
-
-        let error = ensure_collection_write_allowed(&conn, collection_id).unwrap_err();
-        assert!(error.to_string().contains("CollectionRestoringError"));
-        assert!(error.to_string().contains("needs_full_sync=true"));
-    }
-
-    #[test]
-    fn begin_restore_rejects_non_empty_target() {
-        let conn = open_test_db();
-        let source_root = tempfile::TempDir::new().unwrap();
-        let target_root = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", source_root.path());
-        insert_page_with_raw_import(
-            &conn,
-            collection_id,
-            "notes/a",
-            "11111111-1111-7111-8111-111111111111",
-            "hello world from note a",
-            b"---\nmemory_id: 11111111-1111-7111-8111-111111111111\n---\nhello world from note a",
-            "notes/a.md",
-        );
-        fs::write(target_root.path().join("occupied.txt"), b"x").unwrap();
-
-        let error = begin_restore(&conn, "work", target_root.path(), false).unwrap_err();
-        assert!(error.to_string().contains("RestoreNonEmptyTargetError"));
-    }
-
-    #[cfg(not(unix))]
-    #[test]
-    fn begin_restore_on_windows_releases_cli_lease_after_inline_attach_failure() {
-        let (_db_dir, _db_path, conn) = open_test_db_file();
-        let source_root = tempfile::TempDir::new().unwrap();
-        let target_parent = tempfile::TempDir::new().unwrap();
-        let target_root = target_parent.path().join("restored");
-        let collection_id = insert_collection(&conn, "work", source_root.path());
-        let raw_bytes =
-            b"---\nmemory_id: 11111111-1111-7111-8111-111111111111\n---\nhello world from note a";
-        insert_page_with_raw_import(
-            &conn,
-            collection_id,
-            "notes/a",
-            "11111111-1111-7111-8111-111111111111",
-            "hello world from note a",
-            raw_bytes,
-            "notes/a.md",
-        );
-
-        let error = begin_restore(&conn, "work", &target_root, false).unwrap_err();
-
-        assert!(error
-            .to_string()
-            .contains("Vault sync commands require Unix"));
-        type RestoreWindowsFailureRow = (
-            String,
-            String,
-            i64,
-            Option<String>,
-            Option<String>,
-            Option<String>,
-            i64,
-            i64,
-        );
-
-        let row: RestoreWindowsFailureRow = conn
-            .query_row(
-                "SELECT state,
-                        root_path,
-                        needs_full_sync,
-                        pending_root_path,
-                        restore_command_id,
-                        restore_lease_session_id,
-                        (SELECT COUNT(*) FROM collection_owners WHERE collection_id = ?1),
-                        (SELECT COUNT(*) FROM serve_sessions WHERE session_type = 'cli')
-                 FROM collections
-                 WHERE id = ?1",
-                [collection_id],
-                |row| {
-                    Ok((
-                        row.get(0)?,
-                        row.get(1)?,
-                        row.get(2)?,
-                        row.get(3)?,
-                        row.get(4)?,
-                        row.get(5)?,
-                        row.get(6)?,
-                        row.get(7)?,
-                    ))
-                },
-            )
-            .unwrap();
-        assert_eq!(row.0, "restoring");
-        assert_eq!(row.1, target_root.display().to_string());
-        assert_eq!(row.2, 1);
-        assert!(row.3.is_none());
-        assert!(row.4.is_none());
-        assert!(row.5.is_none());
-        assert_eq!(row.6, 0);
-        assert_eq!(row.7, 0);
-        assert_eq!(
-            fs::read(target_root.join("notes").join("a.md")).unwrap(),
-            raw_bytes
-        );
-    }
-
-    #[test]
-    fn mark_collection_restoring_uses_collection_owners_and_clears_ack_residue() {
-        let conn = open_test_db();
-        let temp = tempfile::TempDir::new().unwrap();
-        // Use a high explicit collection id so this test cannot collide with
-        // parallel in-memory tests that share the process-global supervisor registry.
-        let collection_id = insert_collection_with_id(&conn, 50_001, "work", temp.path());
-        conn.execute(
-            "INSERT INTO serve_sessions (session_id, pid, host) VALUES ('serve-owner', 1, 'host')",
-            [],
-        )
-        .unwrap();
-        conn.execute(
-            "INSERT INTO serve_sessions (session_id, pid, host) VALUES ('serve-spoof', 2, 'host')",
-            [],
-        )
-        .unwrap();
-        conn.execute(
-            "INSERT INTO collection_owners (collection_id, session_id) VALUES (?1, 'serve-owner')",
-            [collection_id],
-        )
-        .unwrap();
-        conn.execute(
-            "UPDATE collections
-             SET active_lease_session_id = 'serve-spoof',
-                 reload_generation = 4,
-                 watcher_released_session_id = 'serve-spoof',
-                 watcher_released_generation = 3,
-                 watcher_released_at = strftime('%Y-%m-%dT%H:%M:%SZ', 'now')
-             WHERE id = ?1",
-            [collection_id],
-        )
-        .unwrap();
-
-        let (collection, expected_session_id, generation) =
-            mark_collection_restoring_for_handshake(&conn, collection_id).unwrap();
-
-        assert_eq!(expected_session_id, "serve-owner");
-        assert_eq!(generation, 5);
-        assert_eq!(collection.state, CollectionState::Restoring);
-        assert!(collection.watcher_released_session_id.is_none());
-        assert!(collection.watcher_released_generation.is_none());
-        assert!(collection.watcher_released_at.is_none());
-    }
-
-    // design.md §404-408: mark_collection_restoring_for_handshake must use
-    // live_collection_owner (session_type='serve') so a live CLI lease in
-    // collection_owners is never treated as the expected serve supervisor.
-    #[test]
-    fn mark_collection_restoring_rejects_cli_session_as_handshake_owner() {
-        let conn = open_test_db();
-        let temp = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", temp.path());
-        // Insert a CLI-type session directly into serve_sessions.
-        conn.execute(
-            "INSERT INTO serve_sessions (session_id, pid, host, session_type)
-             VALUES ('cli-lease', 1, 'host', 'cli')",
-            [],
-        )
-        .unwrap();
-        // Force the CLI session as the collection owner (bypasses acquire_owner_lease
-        // type gate to exercise the production handshake path directly).
-        conn.execute(
-            "INSERT INTO collection_owners (collection_id, session_id) VALUES (?1, 'cli-lease')",
-            [collection_id],
-        )
-        .unwrap();
-
-        let err = mark_collection_restoring_for_handshake(&conn, collection_id).unwrap_err();
-
-        // live_collection_owner finds no serve-type owner → ServeOwnsCollectionError,
-        // NOT a timeout waiting for an ack only a serve supervisor can emit.
-        assert!(
-            err.to_string().contains("ServeOwnsCollectionError"),
-            "expected ServeOwnsCollectionError but got: {err}"
-        );
-    }
-
-    // Source-seam invariant: the production handshake paths must use
-    // live_collection_owner (typed) rather than owner_session_id + session_is_live
-    // (untyped).  This guards against regressions that re-open the CLI-as-owner hole.
-    #[test]
     fn handshake_functions_use_typed_live_collection_owner_not_untyped_pair() {
         let src = include_str!("vault_sync.rs");
         // Locate the mark_collection_restoring_for_handshake body (up to its closing
@@ -9183,475 +7887,6 @@ mod tests {
                  use live_collection_owner instead"
             );
         }
-    }
-
-    #[test]
-    fn wait_for_exact_ack_short_circuits_when_live_owner_disappears() {
-        let conn = open_test_db();
-        let temp = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", temp.path());
-        conn.execute(
-            "INSERT INTO serve_sessions (session_id, pid, host) VALUES ('serve-1', 1, 'host')",
-            [],
-        )
-        .unwrap();
-        conn.execute(
-            "INSERT INTO collection_owners (collection_id, session_id) VALUES (?1, 'serve-1')",
-            [collection_id],
-        )
-        .unwrap();
-        conn.execute(
-            "UPDATE collections SET state = 'restoring', reload_generation = 2 WHERE id = ?1",
-            [collection_id],
-        )
-        .unwrap();
-        unregister_session(&conn, "serve-1").unwrap();
-
-        let started = Instant::now();
-        let error = wait_for_exact_ack(&conn, collection_id, "serve-1", 2).unwrap_err();
-
-        assert!(matches!(
-            error,
-            VaultSyncError::ServeDiedDuringHandshake { .. }
-        ));
-        assert!(
-            started.elapsed() < Duration::from_secs(1),
-            "owner-loss path must short-circuit instead of waiting for the full handshake timeout"
-        );
-    }
-
-    #[test]
-    fn acquire_owner_lease_refuses_live_foreign_owner_and_preserves_existing_claim() {
-        let conn = open_test_db();
-        let temp = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", temp.path());
-        conn.execute(
-            "INSERT INTO serve_sessions (session_id, pid, host) VALUES ('serve-owner', 1, 'host')",
-            [],
-        )
-        .unwrap();
-        conn.execute(
-            "INSERT INTO serve_sessions (session_id, pid, host) VALUES ('cli-owner', 2, 'host')",
-            [],
-        )
-        .unwrap();
-        acquire_owner_lease(&conn, collection_id, "serve-owner").unwrap();
-
-        let error = acquire_owner_lease(&conn, collection_id, "cli-owner").unwrap_err();
-
-        assert!(error.to_string().contains("ServeOwnsCollectionError"));
-        assert_eq!(
-            owner_session_id(&conn, collection_id).unwrap().as_deref(),
-            Some("serve-owner")
-        );
-    }
-
-    #[test]
-    fn acquire_owner_lease_allows_same_session_reentrant_claim_and_keeps_single_row() {
-        let conn = open_test_db();
-        let temp = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", temp.path());
-        conn.execute(
-            "INSERT INTO serve_sessions (session_id, pid, host) VALUES ('cli-owner', 2, 'host')",
-            [],
-        )
-        .unwrap();
-
-        acquire_owner_lease(&conn, collection_id, "cli-owner").unwrap();
-        acquire_owner_lease(&conn, collection_id, "cli-owner").unwrap();
-
-        let row: (Option<String>, i64, Option<String>) = conn
-            .query_row(
-                "SELECT active_lease_session_id,
-                        (SELECT COUNT(*) FROM collection_owners WHERE collection_id = ?1),
-                        (SELECT session_id FROM collection_owners WHERE collection_id = ?1)
-                 FROM collections
-                 WHERE id = ?1",
-                [collection_id],
-                |row| Ok((row.get(0)?, row.get(1)?, row.get(2)?)),
-            )
-            .unwrap();
-        assert_eq!(row.0.as_deref(), Some("cli-owner"));
-        assert_eq!(row.1, 1);
-        assert_eq!(row.2.as_deref(), Some("cli-owner"));
-        assert_eq!(
-            owner_session_id(&conn, collection_id).unwrap().as_deref(),
-            Some("cli-owner")
-        );
-    }
-
-    #[test]
-    fn acquire_owner_lease_reclaims_stale_owner_residue_and_updates_mirror_column() {
-        let conn = open_test_db();
-        let temp = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", temp.path());
-        conn.execute(
-            "INSERT INTO serve_sessions (session_id, pid, host, heartbeat_at)
-             VALUES ('stale-owner', 1, 'host', datetime('now', '-120 seconds'))",
-            [],
-        )
-        .unwrap();
-        conn.execute(
-            "INSERT INTO serve_sessions (session_id, pid, host) VALUES ('cli-owner', 2, 'host')",
-            [],
-        )
-        .unwrap();
-        conn.execute(
-            "INSERT INTO collection_owners (collection_id, session_id) VALUES (?1, 'stale-owner')",
-            [collection_id],
-        )
-        .unwrap();
-        conn.execute(
-            "UPDATE collections SET active_lease_session_id = 'stale-owner' WHERE id = ?1",
-            [collection_id],
-        )
-        .unwrap();
-
-        acquire_owner_lease(&conn, collection_id, "cli-owner").unwrap();
-
-        let row: (Option<String>, i64) = conn
-            .query_row(
-                "SELECT active_lease_session_id,
-                        (SELECT COUNT(*) FROM collection_owners WHERE session_id = 'stale-owner')
-                 FROM collections
-                 WHERE id = ?1",
-                [collection_id],
-                |row| Ok((row.get(0)?, row.get(1)?)),
-            )
-            .unwrap();
-        assert_eq!(row.0.as_deref(), Some("cli-owner"));
-        assert_eq!(row.1, 0);
-        assert_eq!(
-            owner_session_id(&conn, collection_id).unwrap().as_deref(),
-            Some("cli-owner")
-        );
-    }
-
-    #[test]
-    fn finalize_pending_restore_requires_exact_originator_or_stale_heartbeat() {
-        let conn = open_test_db();
-        let temp = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", temp.path());
-        conn.execute(
-            "UPDATE collections
-             SET state = 'restoring',
-                 pending_root_path = 'C:/restored',
-                 pending_restore_manifest = '{\"entries\":[]}',
-                 restore_command_id = 'restore-1',
-                 pending_command_heartbeat_at = strftime('%Y-%m-%dT%H:%M:%SZ', 'now')
-             WHERE id = ?1",
-            [collection_id],
-        )
-        .unwrap();
-
-        let outcome = finalize_pending_restore(
-            &conn,
-            collection_id,
-            FinalizeCaller::StartupRecovery {
-                session_id: "serve-1".to_owned(),
-            },
-        )
-        .unwrap();
-        assert_eq!(outcome, FinalizeOutcome::Deferred);
-    }
-
-    #[test]
-    fn finalize_pending_restore_production_callers_pass_explicit_finalize_caller_variants() {
-        let source = production_vault_sync_source();
-        let callsites = source
-            .match_indices("finalize_pending_restore(")
-            .map(|(index, _)| index)
-            .filter(|index| !source[..*index].ends_with("fn "))
-            .collect::<Vec<_>>();
-
-        assert_eq!(
-            callsites.len(),
-            4,
-            "expected exactly four production finalize_pending_restore call sites"
-        );
-
-        for callsite in callsites {
-            let snippet_end = std::cmp::min(callsite + 240, source.len());
-            let snippet = &source[callsite..snippet_end];
-            assert!(
-                snippet.contains("FinalizeCaller::ExternalFinalize")
-                    || snippet.contains("FinalizeCaller::StartupRecovery")
-                    || snippet.contains("FinalizeCaller::RestoreOriginator"),
-                "production finalize call site must pass an explicit FinalizeCaller variant: {snippet}"
-            );
-        }
-    }
-
-    #[test]
-    fn finalize_pending_restore_startup_recovery_uses_shared_15_second_heartbeat_threshold() {
-        let conn = open_test_db();
-        let temp = tempfile::TempDir::new().unwrap();
-        let pending_root = temp.path().join("restored");
-        fs::create_dir_all(&pending_root).unwrap();
-        let collection_id = insert_collection(&conn, "work", temp.path());
-        conn.execute(
-            "UPDATE collections
-             SET state = 'restoring',
-                 pending_root_path = ?2,
-                 pending_restore_manifest = '{\"entries\":[]}',
-                 restore_command_id = 'restore-1',
-                 pending_command_heartbeat_at = datetime('now', '-14 seconds')
-             WHERE id = ?1",
-            params![collection_id, pending_root.display().to_string()],
-        )
-        .unwrap();
-
-        let fresh = finalize_pending_restore(
-            &conn,
-            collection_id,
-            FinalizeCaller::StartupRecovery {
-                session_id: "serve-1".to_owned(),
-            },
-        )
-        .unwrap();
-        assert_eq!(fresh, FinalizeOutcome::Deferred);
-
-        conn.execute(
-            "UPDATE collections
-             SET pending_command_heartbeat_at = datetime('now', '-16 seconds')
-             WHERE id = ?1",
-            [collection_id],
-        )
-        .unwrap();
-
-        let stale = finalize_pending_restore(
-            &conn,
-            collection_id,
-            FinalizeCaller::StartupRecovery {
-                session_id: "serve-1".to_owned(),
-            },
-        )
-        .unwrap();
-        assert_eq!(stale, FinalizeOutcome::Finalized);
-    }
-
-    #[test]
-    fn finalize_pending_restore_allows_exact_originator_with_fresh_heartbeat() {
-        let conn = open_test_db();
-        let temp = tempfile::TempDir::new().unwrap();
-        let pending_root = temp.path().join("restored");
-        fs::create_dir_all(&pending_root).unwrap();
-        let collection_id = insert_collection(&conn, "work", temp.path());
-        conn.execute(
-            "UPDATE collections
-             SET state = 'restoring',
-                 pending_root_path = ?2,
-                 pending_restore_manifest = '{\"entries\":[]}',
-                 restore_command_id = 'restore-1',
-                 pending_command_heartbeat_at = strftime('%Y-%m-%dT%H:%M:%SZ', 'now')
-             WHERE id = ?1",
-            params![collection_id, pending_root.display().to_string()],
-        )
-        .unwrap();
-
-        let outcome = finalize_pending_restore(
-            &conn,
-            collection_id,
-            FinalizeCaller::RestoreOriginator {
-                command_id: "restore-1".to_owned(),
-            },
-        )
-        .unwrap();
-
-        assert_eq!(outcome, FinalizeOutcome::Finalized);
-        let collection = load_collection_by_id(&conn, collection_id).unwrap();
-        assert_eq!(collection.root_path, pending_root.display().to_string());
-        assert!(collection.pending_root_path.is_none());
-        assert!(collection.needs_full_sync);
-    }
-
-    #[test]
-    fn finalize_pending_restore_rejects_foreign_originator_with_fresh_heartbeat() {
-        let conn = open_test_db();
-        let temp = tempfile::TempDir::new().unwrap();
-        let pending_root = temp.path().join("restored");
-        fs::create_dir_all(&pending_root).unwrap();
-        let collection_id = insert_collection(&conn, "work", temp.path());
-        conn.execute(
-            "UPDATE collections
-             SET state = 'restoring',
-                 pending_root_path = ?2,
-                 pending_restore_manifest = '{\"entries\":[]}',
-                 restore_command_id = 'restore-1',
-                 pending_command_heartbeat_at = strftime('%Y-%m-%dT%H:%M:%SZ', 'now')
-             WHERE id = ?1",
-            params![collection_id, pending_root.display().to_string()],
-        )
-        .unwrap();
-
-        let outcome = finalize_pending_restore(
-            &conn,
-            collection_id,
-            FinalizeCaller::RestoreOriginator {
-                command_id: "restore-2".to_owned(),
-            },
-        )
-        .unwrap();
-
-        assert_eq!(outcome, FinalizeOutcome::Deferred);
-        let collection = load_collection_by_id(&conn, collection_id).unwrap();
-        assert_eq!(
-            collection.pending_root_path.as_deref(),
-            Some(pending_root.to_str().unwrap())
-        );
-        assert_eq!(collection.restore_command_id.as_deref(), Some("restore-1"));
-    }
-
-    #[test]
-    fn finalize_pending_restore_external_finalize_runs_tx_b_canonical_state() {
-        let conn = open_test_db();
-        let temp = tempfile::TempDir::new().unwrap();
-        let pending_root = temp.path().join("restored");
-        fs::create_dir_all(&pending_root).unwrap();
-        let collection_id = insert_collection(&conn, "work", temp.path());
-        conn.execute(
-            "UPDATE collections
-             SET state = 'restoring',
-                 pending_root_path = ?2,
-                 pending_restore_manifest = '{\"entries\":[]}',
-                 restore_command_id = 'restore-1',
-                 restore_command_pid = 99,
-                 restore_command_host = 'host',
-                 pending_command_heartbeat_at = datetime('now', '-120 seconds'),
-                 watcher_released_session_id = 'serve-1',
-                 watcher_released_generation = 2,
-                 watcher_released_at = strftime('%Y-%m-%dT%H:%M:%SZ', 'now')
-             WHERE id = ?1",
-            params![collection_id, pending_root.display().to_string()],
-        )
-        .unwrap();
-
-        let outcome = finalize_pending_restore(
-            &conn,
-            collection_id,
-            FinalizeCaller::ExternalFinalize {
-                session_id: "serve-1".to_owned(),
-            },
-        )
-        .unwrap();
-
-        assert_eq!(outcome, FinalizeOutcome::Finalized);
-        let collection = load_collection_by_id(&conn, collection_id).unwrap();
-        assert_eq!(collection.root_path, pending_root.display().to_string());
-        assert_eq!(collection.state, CollectionState::Restoring);
-        assert!(collection.needs_full_sync);
-        assert!(collection.pending_root_path.is_none());
-        assert!(collection.pending_restore_manifest.is_none());
-        assert!(collection.restore_command_id.is_none());
-        assert!(collection.pending_command_heartbeat_at.is_none());
-        assert!(collection.watcher_released_session_id.is_none());
-        assert!(collection.watcher_released_generation.is_none());
-        assert!(collection.watcher_released_at.is_none());
-    }
-
-    #[test]
-    fn run_tx_b_is_idempotent_and_arms_write_gate() {
-        let conn = open_test_db();
-        let temp = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", temp.path());
-        conn.execute(
-            "UPDATE collections
-             SET state = 'restoring',
-                 pending_root_path = ?2,
-                 pending_restore_manifest = '{\"entries\":[]}'
-             WHERE id = ?1",
-            params![collection_id, temp.path().display().to_string()],
-        )
-        .unwrap();
-
-        assert!(run_tx_b(&conn, collection_id).unwrap());
-        assert!(!run_tx_b(&conn, collection_id).unwrap());
-
-        let row: (String, i64) = conn
-            .query_row(
-                "SELECT state, needs_full_sync FROM collections WHERE id = ?1",
-                [collection_id],
-                |row| Ok((row.get(0)?, row.get(1)?)),
-            )
-            .unwrap();
-        assert_eq!(row.0, "restoring");
-        assert_eq!(row.1, 1);
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn remap_online_source_defers_attach_to_rcrt_and_remap_attach_reason_appears_once() {
-        let source = production_vault_sync_source();
-
-        let remap_start = source.find("pub fn remap_collection(").unwrap();
-        let remap_end = source[remap_start..]
-            .find("pub fn verify_remap_root(")
-            .map(|offset| remap_start + offset)
-            .unwrap();
-        let remap_source = &source[remap_start..remap_end];
-        assert!(
-            remap_source.contains(
-                "wait_for_exact_ack(conn, collection.id, &expected_session_id, generation)?"
-            ),
-            "online remap must wait for the exact watcher ack before mutating DB state"
-        );
-        assert!(
-            remap_source.contains("needs_full_sync = 1"),
-            "online remap must arm the write gate for the post-remap attach pass"
-        );
-        assert!(
-            remap_source.contains("DELETE FROM file_state WHERE collection_id = ?1"),
-            "online remap must limit itself to the DB state flip plus file_state reset"
-        );
-        let online_source = &remap_source[..remap_source.find("} else {").unwrap()];
-        assert!(
-            !online_source.contains("complete_attach(")
-                && !online_source.contains("full_hash_reconcile_authorized("),
-            "remap_collection must not run attach or full-hash reconcile inline"
-        );
-
-        let rcrt_start = source.find("pub fn run_rcrt_pass(").unwrap();
-        let rcrt_end = source[rcrt_start..]
-            .find("fn embedding_drain_interval_secs(")
-            .map(|offset| rcrt_start + offset)
-            .unwrap();
-        let rcrt_source = &source[rcrt_start..rcrt_end];
-        assert_eq!(
-            rcrt_source
-                .matches("AttachReason::RemapPostReconcile")
-                .count(),
-            1,
-            "RCRT should have exactly one remap attach arm"
-        );
-        assert!(
-            rcrt_source.contains("if complete_attach(conn, collection_id, session_id, reason)? {"),
-            "RCRT must own the attach transition after remap"
-        );
-    }
-
-    #[test]
-    fn complete_attach_source_is_reentry_guarded_by_needs_full_sync() {
-        let source = production_vault_sync_source();
-        let start = source.find("fn complete_attach(").unwrap();
-        let end = source[start..]
-            .find("pub fn run_rcrt_pass(")
-            .map(|offset| start + offset)
-            .unwrap();
-        let snippet = &source[start..end];
-
-        assert!(
-            snippet.contains("if !collection.needs_full_sync")
-                && snippet.contains("return Ok(false);"),
-            "complete_attach must fail closed to a no-op when the write gate is already cleared"
-        );
-        assert!(
-            snippet.contains("reload_generation = reload_generation + 1"),
-            "attach completion must advance generation exactly on the guarded transition"
-        );
-        assert!(
-            snippet.contains("AND needs_full_sync = 1"),
-            "the attach-completion UPDATE must be gated on needs_full_sync so re-entry cannot bump generation again"
-        );
     }
 
     #[cfg(unix)]
@@ -9744,303 +7979,6 @@ mod tests {
     }
 
     #[test]
-    fn begin_restore_preserves_tx_b_residue_and_plain_sync_cannot_consume_it() {
-        let (_db_dir, _db_path, conn) = open_test_db_file();
-        let source_root = tempfile::TempDir::new().unwrap();
-        let target_parent = tempfile::TempDir::new().unwrap();
-        let target_root = target_parent.path().join("restored");
-        fs::create_dir_all(source_root.path().join("notes")).unwrap();
-        let collection_id = insert_collection(&conn, "work", source_root.path());
-        let raw_bytes =
-            b"---\nmemory_id: 11111111-1111-7111-8111-111111111111\n---\nhello world from note a";
-        insert_page_with_raw_import(
-            &conn,
-            collection_id,
-            "notes/a",
-            "11111111-1111-7111-8111-111111111111",
-            "hello world from note a",
-            raw_bytes,
-            "notes/a.md",
-        );
-        fs::write(source_root.path().join("notes").join("a.md"), raw_bytes).unwrap();
-        conn.execute(
-            "CREATE TRIGGER tx_b_fail
-             BEFORE UPDATE ON collections
-             WHEN OLD.pending_root_path IS NOT NULL AND NEW.pending_root_path IS NULL
-             BEGIN
-                 SELECT RAISE(FAIL, 'tx-b fail');
-             END",
-            [],
-        )
-        .unwrap();
-
-        let error = begin_restore(&conn, "work", &target_root, false).unwrap_err();
-
-        assert!(error.to_string().contains("tx-b fail"));
-        let row: (String, Option<String>, Option<String>, Option<String>) = conn
-            .query_row(
-                "SELECT state, pending_root_path, restore_command_id, pending_restore_manifest
-                 FROM collections
-                 WHERE id = ?1",
-                [collection_id],
-                |row| Ok((row.get(0)?, row.get(1)?, row.get(2)?, row.get(3)?)),
-            )
-            .unwrap();
-        assert_eq!(row.0, "restoring");
-        assert_eq!(row.1.as_deref(), Some(target_root.to_str().unwrap()));
-        assert!(row.2.is_some());
-        assert!(row.3.is_some());
-        assert!(target_root.exists());
-
-        let sync_error = sync_collection(&conn, "work").unwrap_err();
-
-        assert!(sync_error
-            .to_string()
-            .contains("RestorePendingFinalizeError"));
-        let retained_pending_root: Option<String> = conn
-            .query_row(
-                "SELECT pending_root_path FROM collections WHERE id = ?1",
-                [collection_id],
-                |row| row.get(0),
-            )
-            .unwrap();
-        assert_eq!(
-            retained_pending_root.as_deref(),
-            Some(target_root.to_str().unwrap())
-        );
-    }
-
-    #[test]
-    fn finalize_pending_restore_retries_manifest_incomplete_until_success() {
-        let conn = open_test_db();
-        let temp = tempfile::TempDir::new().unwrap();
-        let pending_root = temp.path().join("restored");
-        write_restore_file(&pending_root, "notes/a.md", b"hello from restore");
-        let manifest_json = manifest_json_for_directory(&pending_root);
-        fs::remove_file(pending_root.join("notes").join("a.md")).unwrap();
-        let collection_id = insert_collection(&conn, "work", temp.path());
-        conn.execute(
-            "UPDATE collections
-             SET state = 'restoring',
-                 pending_root_path = ?2,
-                 pending_restore_manifest = ?3,
-                 restore_command_id = 'restore-1',
-                 pending_command_heartbeat_at = strftime('%Y-%m-%dT%H:%M:%SZ', 'now')
-             WHERE id = ?1",
-            params![
-                collection_id,
-                pending_root.display().to_string(),
-                manifest_json
-            ],
-        )
-        .unwrap();
-
-        let first = finalize_pending_restore(
-            &conn,
-            collection_id,
-            FinalizeCaller::RestoreOriginator {
-                command_id: "restore-1".to_owned(),
-            },
-        )
-        .unwrap();
-        assert_eq!(first, FinalizeOutcome::ManifestIncomplete);
-        let first_incomplete_at: String = conn
-            .query_row(
-                "SELECT pending_manifest_incomplete_at FROM collections WHERE id = ?1",
-                [collection_id],
-                |row| row.get(0),
-            )
-            .unwrap();
-
-        let second = finalize_pending_restore(
-            &conn,
-            collection_id,
-            FinalizeCaller::RestoreOriginator {
-                command_id: "restore-1".to_owned(),
-            },
-        )
-        .unwrap();
-        assert_eq!(second, FinalizeOutcome::ManifestIncomplete);
-        let second_incomplete_at: String = conn
-            .query_row(
-                "SELECT pending_manifest_incomplete_at FROM collections WHERE id = ?1",
-                [collection_id],
-                |row| row.get(0),
-            )
-            .unwrap();
-        assert_eq!(first_incomplete_at, second_incomplete_at);
-
-        write_restore_file(&pending_root, "notes/a.md", b"hello from restore");
-        let final_outcome = finalize_pending_restore(
-            &conn,
-            collection_id,
-            FinalizeCaller::RestoreOriginator {
-                command_id: "restore-1".to_owned(),
-            },
-        )
-        .unwrap();
-
-        assert_eq!(final_outcome, FinalizeOutcome::Finalized);
-        let collection = load_collection_by_id(&conn, collection_id).unwrap();
-        assert!(collection.pending_root_path.is_none());
-        assert!(collection.pending_manifest_incomplete_at.is_none());
-        assert!(collection.integrity_failed_at.is_none());
-        assert_eq!(collection.root_path, pending_root.display().to_string());
-    }
-
-    #[test]
-    fn finalize_pending_restore_escalates_manifest_incomplete_after_ttl() {
-        let conn = open_test_db();
-        let temp = tempfile::TempDir::new().unwrap();
-        let pending_root = temp.path().join("restored");
-        write_restore_file(&pending_root, "notes/a.md", b"hello from restore");
-        let manifest_json = manifest_json_for_directory(&pending_root);
-        fs::remove_file(pending_root.join("notes").join("a.md")).unwrap();
-        let collection_id = insert_collection(&conn, "work", temp.path());
-        conn.execute(
-            "UPDATE collections
-             SET state = 'restoring',
-                 pending_root_path = ?2,
-                 pending_restore_manifest = ?3,
-                 restore_command_id = 'restore-1',
-                 pending_command_heartbeat_at = strftime('%Y-%m-%dT%H:%M:%SZ', 'now'),
-                 pending_manifest_incomplete_at = datetime('now', '-31 minutes')
-             WHERE id = ?1",
-            params![
-                collection_id,
-                pending_root.display().to_string(),
-                manifest_json
-            ],
-        )
-        .unwrap();
-
-        let outcome = finalize_pending_restore(
-            &conn,
-            collection_id,
-            FinalizeCaller::RestoreOriginator {
-                command_id: "restore-1".to_owned(),
-            },
-        )
-        .unwrap();
-
-        assert_eq!(outcome, FinalizeOutcome::IntegrityFailed);
-        let collection = load_collection_by_id(&conn, collection_id).unwrap();
-        assert_eq!(
-            collection.pending_root_path.as_deref(),
-            Some(pending_root.to_str().unwrap())
-        );
-        assert!(collection.pending_manifest_incomplete_at.is_some());
-        assert!(collection.integrity_failed_at.is_some());
-    }
-
-    #[test]
-    fn finalize_pending_restore_detects_manifest_tamper_and_restore_reset_clears_it() {
-        let conn = open_test_db();
-        let temp = tempfile::TempDir::new().unwrap();
-        let pending_root = temp.path().join("restored");
-        write_restore_file(&pending_root, "notes/a.md", b"hello from restore");
-        let manifest_json = manifest_json_for_directory(&pending_root);
-        write_restore_file(&pending_root, "notes/a.md", b"tampered bytes");
-        let collection_id = insert_collection(&conn, "work", temp.path());
-        conn.execute(
-            "UPDATE collections
-             SET state = 'restoring',
-                 pending_root_path = ?2,
-                 pending_restore_manifest = ?3,
-                 restore_command_id = 'restore-1',
-                 pending_command_heartbeat_at = strftime('%Y-%m-%dT%H:%M:%SZ', 'now')
-             WHERE id = ?1",
-            params![
-                collection_id,
-                pending_root.display().to_string(),
-                manifest_json
-            ],
-        )
-        .unwrap();
-
-        let outcome = finalize_pending_restore(
-            &conn,
-            collection_id,
-            FinalizeCaller::RestoreOriginator {
-                command_id: "restore-1".to_owned(),
-            },
-        )
-        .unwrap();
-
-        assert_eq!(outcome, FinalizeOutcome::IntegrityFailed);
-        let blocked = load_collection_by_id(&conn, collection_id).unwrap();
-        assert_eq!(
-            blocked.pending_root_path.as_deref(),
-            Some(pending_root.to_str().unwrap())
-        );
-        assert!(blocked.integrity_failed_at.is_some());
-
-        write_restore_file(&pending_root, "notes/a.md", b"hello from restore");
-        let still_blocked = finalize_pending_restore(
-            &conn,
-            collection_id,
-            FinalizeCaller::RestoreOriginator {
-                command_id: "restore-1".to_owned(),
-            },
-        )
-        .unwrap();
-        assert_eq!(still_blocked, FinalizeOutcome::IntegrityFailed);
-
-        restore_reset(&conn, "work").unwrap();
-
-        let reset = load_collection_by_id(&conn, collection_id).unwrap();
-        assert_eq!(reset.state, CollectionState::Active);
-        assert!(reset.pending_root_path.is_none());
-        assert!(reset.pending_restore_manifest.is_none());
-        assert!(reset.restore_command_id.is_none());
-        assert!(reset.integrity_failed_at.is_none());
-    }
-
-    #[test]
-    fn unregister_session_clears_ownership_mirror_columns() {
-        let conn = open_test_db();
-        let temp = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", temp.path());
-        conn.execute(
-            "INSERT INTO serve_sessions (session_id, pid, host) VALUES ('serve-1', 1, 'host')",
-            [],
-        )
-        .unwrap();
-        conn.execute(
-            "INSERT INTO collection_owners (collection_id, session_id) VALUES (?1, 'serve-1')",
-            [collection_id],
-        )
-        .unwrap();
-        conn.execute(
-            "UPDATE collections
-             SET active_lease_session_id = 'serve-1',
-                 restore_lease_session_id = 'serve-1'
-             WHERE id = ?1",
-            [collection_id],
-        )
-        .unwrap();
-
-        unregister_session(&conn, "serve-1").unwrap();
-
-        let row: (Option<String>, Option<String>, i64, i64) = conn
-            .query_row(
-                "SELECT active_lease_session_id,
-                        restore_lease_session_id,
-                        (SELECT COUNT(*) FROM collection_owners WHERE session_id = 'serve-1'),
-                        (SELECT COUNT(*) FROM serve_sessions WHERE session_id = 'serve-1')
-                 FROM collections
-                 WHERE id = ?1",
-                [collection_id],
-                |row| Ok((row.get(0)?, row.get(1)?, row.get(2)?, row.get(3)?)),
-            )
-            .unwrap();
-        assert!(row.0.is_none());
-        assert!(row.1.is_none());
-        assert_eq!(row.2, 0);
-        assert_eq!(row.3, 0);
-    }
-
-    #[test]
     fn lease_guard_drop_releases_owner_lease_and_unregisters_short_lived_session() {
         let (_dir, db_path, conn) = open_test_db_file();
         let temp = tempfile::TempDir::new().unwrap();
@@ -10070,599 +8008,6 @@ mod tests {
         assert!(row.0.is_none());
         assert_eq!(row.1, 0);
         assert_eq!(row.2, 0);
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn offline_restore_runs_attach_inline_and_reopens_writes() {
-        let (_db_dir, _db_path, conn) = open_test_db_file();
-        let source_root = tempfile::TempDir::new().unwrap();
-        let target_parent = tempfile::TempDir::new().unwrap();
-        let target_root = target_parent.path().join("restored");
-        fs::create_dir_all(source_root.path().join("notes")).unwrap();
-        let collection_id = insert_collection(&conn, "work", source_root.path());
-        let raw_bytes =
-            b"---\nmemory_id: 11111111-1111-7111-8111-111111111111\n---\nhello world from note a";
-        insert_page_with_raw_import(
-            &conn,
-            collection_id,
-            "notes/a",
-            "11111111-1111-7111-8111-111111111111",
-            "hello world from note a",
-            raw_bytes,
-            "notes/a.md",
-        );
-        fs::write(source_root.path().join("notes").join("a.md"), raw_bytes).unwrap();
-
-        begin_restore(&conn, "work", &target_root, false).unwrap();
-
-        let collection = load_collection_by_id(&conn, collection_id).unwrap();
-        assert_eq!(collection.state, CollectionState::Active);
-        assert!(!collection.needs_full_sync);
-        assert!(collection.active_lease_session_id.is_none());
-        assert!(collection.restore_lease_session_id.is_none());
-        assert!(owner_session_id(&conn, collection_id).unwrap().is_none());
-        ensure_collection_write_allowed(&conn, collection_id).unwrap();
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn offline_remap_runs_reconcile_inline_and_preserves_uuid_identity_across_reorganization() {
-        let (_db_dir, _db_path, conn) = open_test_db_file();
-        let old_root = tempfile::TempDir::new().unwrap();
-        let new_root = tempfile::TempDir::new().unwrap();
-        fs::create_dir_all(old_root.path().join("notes")).unwrap();
-        let collection_id = insert_collection(&conn, "work", old_root.path());
-        let raw_bytes_a =
-            b"---\nmemory_id: 11111111-1111-7111-8111-111111111111\nslug: notes/a\n---\nhello world from note a";
-        let raw_bytes_b =
-            b"---\nmemory_id: 22222222-2222-7222-8222-222222222222\nslug: notes/b\n---\nhello world from note b";
-        let page_a = insert_page_with_raw_import(
-            &conn,
-            collection_id,
-            "notes/a",
-            "11111111-1111-7111-8111-111111111111",
-            "hello world from note a",
-            raw_bytes_a,
-            "notes/old-a.md",
-        );
-        let page_b = insert_page_with_raw_import(
-            &conn,
-            collection_id,
-            "notes/b",
-            "22222222-2222-7222-8222-222222222222",
-            "hello world from note b",
-            raw_bytes_b,
-            "notes/b.md",
-        );
-        fs::write(old_root.path().join("notes").join("old-a.md"), raw_bytes_a).unwrap();
-        fs::write(old_root.path().join("notes").join("b.md"), raw_bytes_b).unwrap();
-        fs::create_dir_all(new_root.path().join("notes")).unwrap();
-        fs::create_dir_all(new_root.path().join("nested")).unwrap();
-        fs::write(
-            new_root.path().join("nested").join("renamed-a.md"),
-            raw_bytes_a,
-        )
-        .unwrap();
-        fs::write(new_root.path().join("notes").join("b.md"), raw_bytes_b).unwrap();
-        conn.execute(
-            "INSERT INTO links (from_page_id, to_page_id, relationship, context, source_kind)
-             VALUES (?1, ?2, 'depends_on', '', 'programmatic')",
-            params![page_a, page_b],
-        )
-        .unwrap();
-
-        let summary = remap_collection(&conn, "work", new_root.path(), false).unwrap();
-
-        let collection = load_collection_by_id(&conn, collection_id).unwrap();
-        assert_eq!(collection.root_path, new_root.path().display().to_string());
-        assert_eq!(collection.state, CollectionState::Active);
-        assert!(!collection.needs_full_sync);
-        assert_eq!(summary.resolved_pages, 2);
-        assert!(collection.active_lease_session_id.is_none());
-        assert!(collection.restore_lease_session_id.is_none());
-        assert!(owner_session_id(&conn, collection_id).unwrap().is_none());
-        ensure_collection_write_allowed(&conn, collection_id).unwrap();
-        let remapped_page_id: i64 = conn
-            .query_row(
-                "SELECT id FROM pages WHERE collection_id = ?1 AND slug = 'notes/a'",
-                [collection_id],
-                |row| row.get(0),
-            )
-            .unwrap();
-        assert_eq!(remapped_page_id, page_a);
-        let relative_path: String = conn
-            .query_row(
-                "SELECT relative_path FROM file_state WHERE page_id = ?1",
-                [page_a],
-                |row| row.get(0),
-            )
-            .unwrap();
-        assert_eq!(relative_path, "nested/renamed-a.md");
-        let link_count: i64 = conn
-            .query_row(
-                "SELECT COUNT(*) FROM links WHERE from_page_id = ?1 AND to_page_id = ?2",
-                params![page_a, page_b],
-                |row| row.get(0),
-            )
-            .unwrap();
-        assert_eq!(link_count, 1);
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn remap_collection_refuses_phase1_drift_until_new_root_catches_up() {
-        let (_db_dir, _db_path, conn) = open_test_db_file();
-        let old_root = tempfile::TempDir::new().unwrap();
-        let new_root = tempfile::TempDir::new().unwrap();
-        fs::create_dir_all(old_root.path().join("notes")).unwrap();
-        let collection_id = insert_collection(&conn, "work", old_root.path());
-        let original =
-            b"---\nmemory_id: 11111111-1111-7111-8111-111111111111\n---\nhello world from note a";
-        insert_page_with_raw_import(
-            &conn,
-            collection_id,
-            "notes/a",
-            "11111111-1111-7111-8111-111111111111",
-            "hello world from note a",
-            original,
-            "notes/a.md",
-        );
-        fs::write(old_root.path().join("notes").join("a.md"), original).unwrap();
-        fs::create_dir_all(new_root.path().join("notes")).unwrap();
-        fs::write(new_root.path().join("notes").join("a.md"), original).unwrap();
-
-        let updated =
-            b"---\nmemory_id: 11111111-1111-7111-8111-111111111111\n---\nupdated body before remap";
-        fs::write(old_root.path().join("notes").join("a.md"), updated).unwrap();
-
-        let first = remap_collection(&conn, "work", new_root.path(), false).unwrap_err();
-        assert!(first.to_string().contains("RemapDriftConflictError"));
-        let active_raw_import: Vec<u8> = conn
-            .query_row(
-                "SELECT raw_bytes FROM raw_imports WHERE page_id = (SELECT id FROM pages WHERE collection_id = ?1 AND slug = 'notes/a') AND is_active = 1",
-                [collection_id],
-                |row| row.get(0),
-            )
-            .unwrap();
-        assert_eq!(active_raw_import, updated);
-
-        fs::write(new_root.path().join("notes").join("a.md"), updated).unwrap();
-        let second = remap_collection(&conn, "work", new_root.path(), false).unwrap();
-        assert_eq!(second.missing_pages, 0);
-        assert_eq!(second.mismatched_pages, 0);
-        assert_eq!(second.extra_files, 0);
-    }
-
-    #[cfg(not(unix))]
-    #[test]
-    fn remap_collection_fails_closed_on_windows_before_mutating_collection_state() {
-        let (_db_dir, _db_path, conn) = open_test_db_file();
-        let old_root = tempfile::TempDir::new().unwrap();
-        let new_root = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", old_root.path());
-        let original_root = old_root.path().display().to_string();
-
-        let error = remap_collection(&conn, "work", new_root.path(), false).unwrap_err();
-
-        assert!(error
-            .to_string()
-            .contains("Vault sync commands require Unix"));
-        let row: (String, String, i64) = conn
-            .query_row(
-                "SELECT root_path, state, needs_full_sync FROM collections WHERE id = ?1",
-                [collection_id],
-                |row| Ok((row.get(0)?, row.get(1)?, row.get(2)?)),
-            )
-            .unwrap();
-        assert_eq!(row.0, original_root);
-        assert_eq!(row.1, "active");
-        assert_eq!(row.2, 0);
-    }
-
-    #[test]
-    fn verify_remap_root_uses_unique_hash_fallback_and_ignores_quaidignore_patterns() {
-        let conn = open_test_db();
-        let old_root = tempfile::TempDir::new().unwrap();
-        let new_root = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", old_root.path());
-        let raw_bytes = b"---\ntitle: Hash Fallback\ntype: concept\n---\nthis body is intentionally long enough to cross the remap hash fallback threshold.\n";
-        insert_page_with_raw_import(
-            &conn,
-            collection_id,
-            "notes/hash-fallback",
-            "11111111-1111-7111-8111-111111111111",
-            "this body is intentionally long enough to cross the remap hash fallback threshold.",
-            raw_bytes,
-            "notes/hash-fallback.md",
-        );
-        fs::create_dir_all(new_root.path().join("nested")).unwrap();
-        fs::write(new_root.path().join("nested").join("moved.md"), raw_bytes).unwrap();
-        fs::write(new_root.path().join(".quaidignore"), "ignored/**\n").unwrap();
-        fs::create_dir_all(new_root.path().join("ignored")).unwrap();
-        fs::write(
-            new_root.path().join("ignored").join("secret.md"),
-            b"top secret",
-        )
-        .unwrap();
-
-        let collection = load_collection_by_id(&conn, collection_id).unwrap();
-        let summary = verify_remap_root(&conn, &collection, new_root.path()).unwrap();
-
-        assert_eq!(summary.resolved_pages, 1);
-        assert_eq!(summary.missing_pages, 0);
-        assert_eq!(summary.mismatched_pages, 0);
-        assert_eq!(summary.extra_files, 0);
-    }
-
-    #[test]
-    fn verify_remap_root_ignores_non_markdown_files() {
-        let conn = open_test_db();
-        let old_root = tempfile::TempDir::new().unwrap();
-        let new_root = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", old_root.path());
-        let raw_bytes =
-            b"---\nmemory_id: 11111111-1111-7111-8111-111111111111\n---\nhello world from note a";
-        insert_page_with_raw_import(
-            &conn,
-            collection_id,
-            "notes/a",
-            "11111111-1111-7111-8111-111111111111",
-            "hello world from note a",
-            raw_bytes,
-            "notes/a.md",
-        );
-        fs::create_dir_all(new_root.path().join("notes")).unwrap();
-        fs::create_dir_all(new_root.path().join("assets")).unwrap();
-        fs::write(new_root.path().join("notes").join("a.md"), raw_bytes).unwrap();
-        fs::write(
-            new_root.path().join("assets").join("logo.png"),
-            b"not markdown, not a remap extra",
-        )
-        .unwrap();
-
-        let collection = load_collection_by_id(&conn, collection_id).unwrap();
-        let summary = verify_remap_root(&conn, &collection, new_root.path()).unwrap();
-
-        assert_eq!(summary.resolved_pages, 1);
-        assert_eq!(summary.missing_pages, 0);
-        assert_eq!(summary.mismatched_pages, 0);
-        assert_eq!(summary.extra_files, 0);
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn verify_remap_root_skips_symlinked_entries_in_new_root() {
-        use std::os::unix::fs::symlink;
-
-        let conn = open_test_db();
-        let old_root = tempfile::TempDir::new().unwrap();
-        let new_root = tempfile::TempDir::new().unwrap();
-        let linked = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", old_root.path());
-        let raw_bytes =
-            b"---\nmemory_id: 11111111-1111-7111-8111-111111111111\n---\nhello world from note a";
-        insert_page_with_raw_import(
-            &conn,
-            collection_id,
-            "notes/a",
-            "11111111-1111-7111-8111-111111111111",
-            "hello world from note a",
-            raw_bytes,
-            "notes/a.md",
-        );
-        fs::create_dir_all(new_root.path().join("notes")).unwrap();
-        fs::write(new_root.path().join("notes").join("a.md"), raw_bytes).unwrap();
-        fs::create_dir_all(linked.path().join("shadow")).unwrap();
-        fs::write(
-            linked.path().join("shadow").join("extra.md"),
-            b"reachable only through symlink",
-        )
-        .unwrap();
-        symlink(
-            linked.path().join("shadow"),
-            new_root.path().join("linked-shadow"),
-        )
-        .unwrap();
-
-        let collection = load_collection_by_id(&conn, collection_id).unwrap();
-        let summary = verify_remap_root(&conn, &collection, new_root.path()).unwrap();
-
-        assert_eq!(summary.resolved_pages, 1);
-        assert_eq!(summary.missing_pages, 0);
-        assert_eq!(summary.mismatched_pages, 0);
-        assert_eq!(summary.extra_files, 0);
-    }
-
-    #[test]
-    fn verify_remap_root_rejects_invalid_frontmatter_uuid_in_new_root() {
-        let conn = open_test_db();
-        let old_root = tempfile::TempDir::new().unwrap();
-        let new_root = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", old_root.path());
-        insert_page_with_raw_import(
-            &conn,
-            collection_id,
-            "notes/a",
-            "11111111-1111-7111-8111-111111111111",
-            "hello world from note a",
-            b"---\nmemory_id: 11111111-1111-7111-8111-111111111111\n---\nhello world from note a",
-            "notes/a.md",
-        );
-        fs::create_dir_all(new_root.path().join("notes")).unwrap();
-        fs::write(
-            new_root.path().join("notes").join("broken.md"),
-            b"---\nmemory_id: definitely-not-a-uuid\n---\nhello world from note a",
-        )
-        .unwrap();
-
-        let collection = load_collection_by_id(&conn, collection_id).unwrap();
-        let error = verify_remap_root(&conn, &collection, new_root.path()).unwrap_err();
-
-        assert!(matches!(error, VaultSyncError::InvariantViolation { .. }));
-        assert!(error.to_string().contains("invalid"));
-    }
-
-    #[test]
-    fn verify_remap_root_does_not_use_hash_fallback_for_short_body() {
-        let conn = open_test_db();
-        let old_root = tempfile::TempDir::new().unwrap();
-        let new_root = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", old_root.path());
-        let raw_bytes = b"---\ntitle: Short\n---\nshort\n";
-        insert_page_with_raw_import(
-            &conn,
-            collection_id,
-            "notes/short",
-            "11111111-1111-7111-8111-111111111111",
-            "short",
-            raw_bytes,
-            "notes/short.md",
-        );
-        fs::create_dir_all(new_root.path().join("notes")).unwrap();
-        fs::write(new_root.path().join("notes").join("moved.md"), raw_bytes).unwrap();
-
-        let collection = load_collection_by_id(&conn, collection_id).unwrap();
-        let error = verify_remap_root(&conn, &collection, new_root.path()).unwrap_err();
-
-        assert!(matches!(
-            error,
-            VaultSyncError::NewRootVerificationFailed {
-                missing: 1,
-                mismatched: 0,
-                extra: 1,
-                ..
-            }
-        ));
-    }
-
-    #[test]
-    fn verify_remap_root_rejects_duplicate_hash_candidates_without_uuid_match() {
-        let conn = open_test_db();
-        let old_root = tempfile::TempDir::new().unwrap();
-        let new_root = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", old_root.path());
-        let raw_bytes = b"---\ntitle: Duplicate Hash\n---\nthis body is intentionally long enough to cross the remap hash fallback threshold.\n";
-        insert_page_with_raw_import(
-            &conn,
-            collection_id,
-            "notes/hash-duplicate",
-            "11111111-1111-7111-8111-111111111111",
-            "this body is intentionally long enough to cross the remap hash fallback threshold.",
-            raw_bytes,
-            "notes/hash-duplicate.md",
-        );
-        fs::create_dir_all(new_root.path().join("notes")).unwrap();
-        fs::write(new_root.path().join("notes").join("one.md"), raw_bytes).unwrap();
-        fs::write(new_root.path().join("notes").join("two.md"), raw_bytes).unwrap();
-
-        let collection = load_collection_by_id(&conn, collection_id).unwrap();
-        let error = verify_remap_root(&conn, &collection, new_root.path()).unwrap_err();
-
-        assert!(matches!(
-            error,
-            VaultSyncError::NewRootVerificationFailed {
-                missing: 1,
-                mismatched: 0,
-                extra: 2,
-                ..
-            }
-        ));
-    }
-
-    #[test]
-    fn verify_remap_root_reports_missing_and_extra_counts() {
-        let conn = open_test_db();
-        let old_root = tempfile::TempDir::new().unwrap();
-        let new_root = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", old_root.path());
-        insert_page_with_raw_import(
-            &conn,
-            collection_id,
-            "notes/a",
-            "11111111-1111-7111-8111-111111111111",
-            "hello world from note a",
-            b"---\nmemory_id: 11111111-1111-7111-8111-111111111111\n---\nhello world from note a",
-            "notes/a.md",
-        );
-        insert_page_with_raw_import(
-            &conn,
-            collection_id,
-            "notes/b",
-            "22222222-2222-7222-8222-222222222222",
-            "hello world from note b",
-            b"---\nmemory_id: 22222222-2222-7222-8222-222222222222\n---\nhello world from note b",
-            "notes/b.md",
-        );
-        fs::create_dir_all(new_root.path().join("notes")).unwrap();
-        fs::write(
-            new_root.path().join("notes").join("a.md"),
-            b"---\nmemory_id: 11111111-1111-7111-8111-111111111111\n---\nhello world from note a",
-        )
-        .unwrap();
-        fs::write(
-            new_root.path().join("notes").join("extra.md"),
-            b"extra file",
-        )
-        .unwrap();
-
-        let collection = load_collection_by_id(&conn, collection_id).unwrap();
-        let error = verify_remap_root(&conn, &collection, new_root.path()).unwrap_err();
-        assert!(matches!(
-            error,
-            VaultSyncError::NewRootVerificationFailed {
-                missing: 1,
-                mismatched: 0,
-                extra: 1,
-                ..
-            }
-        ));
-    }
-
-    #[test]
-    fn verify_remap_root_error_includes_sampled_diffs() {
-        let conn = open_test_db();
-        let old_root = tempfile::TempDir::new().unwrap();
-        let new_root = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", old_root.path());
-        insert_page_with_raw_import(
-            &conn,
-            collection_id,
-            "notes/a",
-            "11111111-1111-7111-8111-111111111111",
-            "hello world from note a",
-            b"---\nmemory_id: 11111111-1111-7111-8111-111111111111\n---\nhello world from note a",
-            "notes/a.md",
-        );
-        insert_page_with_raw_import(
-            &conn,
-            collection_id,
-            "notes/b",
-            "22222222-2222-7222-8222-222222222222",
-            "hello world from note b",
-            b"---\nmemory_id: 22222222-2222-7222-8222-222222222222\n---\nhello world from note b",
-            "notes/b.md",
-        );
-        fs::create_dir_all(new_root.path().join("notes")).unwrap();
-        fs::write(
-            new_root.path().join("notes").join("a.md"),
-            b"---\nmemory_id: 11111111-1111-7111-8111-111111111111\n---\nchanged bytes",
-        )
-        .unwrap();
-        fs::write(
-            new_root.path().join("notes").join("extra.md"),
-            b"extra file",
-        )
-        .unwrap();
-
-        let collection = load_collection_by_id(&conn, collection_id).unwrap();
-        let error = verify_remap_root(&conn, &collection, new_root.path()).unwrap_err();
-        let rendered = error.to_string();
-
-        assert!(rendered.contains("missing_samples=notes/b"));
-        assert!(rendered.contains("mismatched_samples=notes/a -> notes/a.md sha256_mismatch"));
-        assert!(rendered.contains("extra_samples=notes/extra.md"));
-    }
-
-    #[test]
-    fn verify_remap_root_reports_mismatched_count_for_duplicate_uuid_candidates() {
-        let conn = open_test_db();
-        let old_root = tempfile::TempDir::new().unwrap();
-        let new_root = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", old_root.path());
-        insert_page_with_raw_import(
-            &conn,
-            collection_id,
-            "notes/a",
-            "11111111-1111-7111-8111-111111111111",
-            "hello world from note a",
-            b"---\nmemory_id: 11111111-1111-7111-8111-111111111111\n---\nhello world from note a",
-            "notes/a.md",
-        );
-        fs::create_dir_all(new_root.path().join("notes")).unwrap();
-        fs::write(
-            new_root.path().join("notes").join("a-one.md"),
-            b"---\nmemory_id: 11111111-1111-7111-8111-111111111111\n---\nfirst duplicate",
-        )
-        .unwrap();
-        fs::write(
-            new_root.path().join("notes").join("a-two.md"),
-            b"---\nmemory_id: 11111111-1111-7111-8111-111111111111\n---\nsecond duplicate",
-        )
-        .unwrap();
-
-        let collection = load_collection_by_id(&conn, collection_id).unwrap();
-        let error = verify_remap_root(&conn, &collection, new_root.path()).unwrap_err();
-        assert!(matches!(
-            error,
-            VaultSyncError::NewRootVerificationFailed {
-                missing: 1,
-                mismatched: 1,
-                extra: 2,
-                ..
-            }
-        ));
-    }
-
-    #[test]
-    fn verify_remap_root_source_uses_before_and_after_tree_fences() {
-        let source = production_vault_sync_source();
-        let start = source.find("pub fn verify_remap_root(").unwrap();
-        let end = source[start..]
-            .find("pub fn restore_reset(")
-            .map(|offset| start + offset)
-            .unwrap();
-        let snippet = &source[start..end];
-
-        assert!(
-            snippet.contains("let before = take_tree_fence(new_root)?;")
-                && snippet.contains("let after = take_tree_fence(new_root)?;"),
-            "verify_remap_root must fence the tree before and after matching to catch mid-flight drift"
-        );
-        assert!(
-            snippet.contains("return Err(VaultSyncError::NewRootUnstable"),
-            "verify_remap_root must fail closed with NewRootUnstableError when the fence changes"
-        );
-    }
-
-    #[test]
-    fn load_new_root_files_source_applies_new_root_ignore_matcher() {
-        let source = production_vault_sync_source();
-        let start = source.find("fn load_new_root_files(").unwrap();
-        let end = source[start..]
-            .find("fn resolve_page_matches(")
-            .map(|offset| start + offset)
-            .unwrap();
-        let snippet = &source[start..end];
-
-        assert!(
-            snippet.contains("let ignore_globset = build_new_root_ignore_globset(root)?;"),
-            "load_new_root_files must build a fresh ignore matcher from the remap root before counting files"
-        );
-        assert!(
-            snippet.contains("if ignore_globset.is_match(relative_path) {"),
-            "load_new_root_files must exclude .quaidignore-matched files from remap verification counts"
-        );
-        assert!(
-            snippet.contains("if !is_markdown_file(relative_path) {"),
-            "load_new_root_files must reuse the reconciler's Markdown gate so Phase 4 extra counts stay in parity"
-        );
-    }
-
-    #[test]
-    fn resolve_page_matches_source_uses_canonical_resolver_helper() {
-        let source = production_vault_sync_source();
-        let start = source.find("fn resolve_page_matches(").unwrap();
-        let end = source[start..]
-            .find("fn take_tree_fence(")
-            .map(|offset| start + offset)
-            .unwrap();
-        let snippet = &source[start..end];
-
-        assert!(
-            snippet.contains("resolve_page_identity("),
-            "Phase 4 remap verification must invoke the canonical resolve_page_identity helper instead of bespoke matching"
-        );
     }
 
     #[test]
@@ -10763,227 +8108,6 @@ mod tests {
     }
 
     #[test]
-    fn take_tree_fence_source_uses_full_stat_tuple() {
-        let source = production_vault_sync_source();
-        assert!(
-            source.contains("metadata_timestamp_ns(metadata.mtime(), metadata.mtime_nsec())")
-                && source.contains("metadata_timestamp_ns(metadata.ctime(), metadata.ctime_nsec())")
-                && source.contains("metadata.ino()"),
-            "Phase 4 tree fence must capture the full per-file stat tuple so same-size rewrites and atomic replacements cannot slip past remap verification"
-        );
-    }
-
-    #[test]
-    fn walk_tree_source_skips_symlinked_entries() {
-        let source = production_vault_sync_source();
-        let start = source.find("fn walk_tree(").unwrap();
-        let end = source[start..]
-            .find("fn path_string(")
-            .map(|offset| start + offset)
-            .unwrap();
-        let snippet = &source[start..end];
-
-        assert!(
-            snippet.contains("let file_type = entry.file_type()?;")
-                && snippet.contains("if file_type.is_symlink() {"),
-            "Phase 4 tree walks must inspect entry file types without following symlinks"
-        );
-    }
-
-    #[test]
-    fn remap_source_runs_safety_pipeline_before_new_root_verification() {
-        let source = production_vault_sync_source();
-        let remap_start = source.find("pub fn remap_collection(").unwrap();
-        let remap_end = source[remap_start..]
-            .find("pub fn verify_remap_root(")
-            .map(|offset| remap_start + offset)
-            .unwrap();
-        let remap_source = &source[remap_start..remap_end];
-
-        let safety_idx = remap_source
-            .find("run_restore_remap_safety_pipeline_without_mount_check")
-            .unwrap();
-        let verify_idx = remap_source
-            .find("verify_remap_root(conn, &collection, new_root)?")
-            .unwrap();
-        assert!(
-            safety_idx < verify_idx,
-            "remap_collection must capture old-root drift before trusting the new root"
-        );
-    }
-
-    #[test]
-    fn remap_online_source_waits_for_exact_ack_before_safety_pipeline() {
-        let source = production_vault_sync_source();
-        let remap_start = source.find("pub fn remap_collection(").unwrap();
-        let remap_end = source[remap_start..]
-            .find("pub fn verify_remap_root(")
-            .map(|offset| remap_start + offset)
-            .unwrap();
-        let remap_source = &source[remap_start..remap_end];
-        let ack_idx = remap_source
-            .find("wait_for_exact_ack(conn, collection.id, &expected_session_id, generation)?")
-            .unwrap();
-        let safety_idx = remap_source
-            .find("run_restore_remap_safety_pipeline_without_mount_check")
-            .unwrap();
-
-        assert!(
-            ack_idx < safety_idx,
-            "online remap must release the live watcher and then capture old-root drift under the acknowledged owner lease"
-        );
-    }
-
-    #[test]
-    fn restore_source_runs_safety_pipeline_before_materialization() {
-        let source = production_vault_sync_source();
-        let restore_start = source.find("pub fn begin_restore(").unwrap();
-        let restore_end = source[restore_start..]
-            .find("pub fn remap_collection(")
-            .map(|offset| restore_start + offset)
-            .unwrap();
-        let restore_source = &source[restore_start..restore_end];
-
-        let online_start = restore_source
-            .find("let (_, expected_session_id, generation) =")
-            .unwrap();
-        let online_end = restore_source[online_start..]
-            .find("} else {")
-            .map(|offset| online_start + offset)
-            .unwrap();
-        let online_source = &restore_source[online_start..online_end];
-        let online_safety_idx = online_source
-            .find("run_restore_remap_safety_pipeline_without_mount_check")
-            .unwrap();
-        let online_materialize_idx = online_source
-            .find("materialize_collection_to_path(conn, &collection, &staging_path)?;")
-            .unwrap();
-        assert!(
-            online_safety_idx < online_materialize_idx,
-            "online restore must capture old-root drift before materializing raw_imports into the target"
-        );
-
-        let offline_start = restore_source
-            .find("let lease = start_short_lived_owner_lease(conn, collection.id)?;")
-            .unwrap();
-        let offline_source = &restore_source[offline_start..];
-        let offline_safety_idx = offline_source
-            .find("run_restore_remap_safety_pipeline_without_mount_check")
-            .unwrap();
-        let offline_materialize_idx = offline_source
-            .find("materialize_collection_to_path(conn, &collection, &staging_path)?;")
-            .unwrap();
-        assert!(
-            offline_safety_idx < offline_materialize_idx,
-            "offline restore must capture old-root drift before materializing raw_imports into the target"
-        );
-    }
-
-    #[test]
-    fn restore_online_source_waits_for_exact_ack_before_safety_pipeline() {
-        let source = production_vault_sync_source();
-        let restore_start = source.find("pub fn begin_restore(").unwrap();
-        let restore_end = source[restore_start..]
-            .find("pub fn remap_collection(")
-            .map(|offset| restore_start + offset)
-            .unwrap();
-        let restore_source = &source[restore_start..restore_end];
-        let online_start = restore_source
-            .find("let (_, expected_session_id, generation) =")
-            .unwrap();
-        let online_end = restore_source[online_start..]
-            .find("} else {")
-            .map(|offset| online_start + offset)
-            .unwrap();
-        let online_source = &restore_source[online_start..online_end];
-        let ack_idx = online_source
-            .find("wait_for_exact_ack(conn, collection.id, &expected_session_id, generation)?")
-            .unwrap();
-        let safety_idx = online_source
-            .find("run_restore_remap_safety_pipeline_without_mount_check")
-            .unwrap();
-
-        assert!(
-            ack_idx < safety_idx,
-            "online restore must wait for the acknowledged owner lease before drift capture"
-        );
-    }
-
-    #[test]
-    fn remap_source_verifies_new_root_before_switching_root_path() {
-        let source = production_vault_sync_source();
-        let remap_start = source.find("pub fn remap_collection(").unwrap();
-        let remap_end = source[remap_start..]
-            .find("pub fn verify_remap_root(")
-            .map(|offset| remap_start + offset)
-            .unwrap();
-        let remap_source = &source[remap_start..remap_end];
-        let verify_idx = remap_source
-            .find("verify_remap_root(conn, &collection, new_root)?")
-            .unwrap();
-        let update_idx = remap_source.find("SET root_path = ?2,").unwrap();
-
-        assert!(
-            verify_idx < update_idx,
-            "remap_collection must prove the target tree before rewriting the collection root"
-        );
-    }
-
-    #[test]
-    fn remap_offline_source_uses_short_lived_lease_and_inline_attach() {
-        let source = production_vault_sync_source();
-        let remap_start = source.find("pub fn remap_collection(").unwrap();
-        let remap_end = source[remap_start..]
-            .find("pub fn verify_remap_root(")
-            .map(|offset| remap_start + offset)
-            .unwrap();
-        let remap_source = &source[remap_start..remap_end];
-        let offline_start = remap_source
-            .find("let lease = start_short_lived_owner_lease")
-            .unwrap();
-        let offline_end = remap_source[offline_start..]
-            .find("Ok(summary)")
-            .map(|offset| offline_start + offset)
-            .unwrap();
-        let offline_source = &remap_source[offline_start..offline_end];
-
-        assert!(
-            offline_source.contains("complete_attach(")
-                && offline_source.contains("AttachReason::RemapPostReconcile"),
-            "offline remap must run the attach/full-hash path inline while the CLI lease is live"
-        );
-        assert!(
-            !offline_source.contains("unregister_session("),
-            "offline remap must not drop its lease before the inline attach finishes"
-        );
-    }
-
-    #[test]
-    fn restore_offline_source_uses_short_lived_lease_and_inline_attach() {
-        let source = production_vault_sync_source();
-        let restore_start = source.find("pub fn begin_restore(").unwrap();
-        let restore_end = source[restore_start..]
-            .find("pub fn remap_collection(")
-            .map(|offset| restore_start + offset)
-            .unwrap();
-        let restore_source = &source[restore_start..restore_end];
-        let offline_start = restore_source
-            .find("let lease = start_short_lived_owner_lease(conn, collection.id)?;")
-            .unwrap();
-        let offline_source = &restore_source[offline_start..];
-
-        assert!(
-            offline_source.contains("complete_attach(")
-                && offline_source.contains("AttachReason::RestorePostFinalize"),
-            "offline restore must run the attach/full-hash path inline while the CLI lease is live"
-        );
-        assert!(
-            !offline_source.contains("unregister_session("),
-            "offline restore must not drop its lease before the inline attach finishes"
-        );
-    }
-
-    #[test]
     fn run_rcrt_pass_skips_reconcile_halted_collections() {
         const COLLECTION_ID: i64 = 50_002;
 
@@ -11016,259 +8140,6 @@ mod tests {
             actions,
             vec![(collection_id, "work:skipped-halt".to_owned())]
         );
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn start_serve_runtime_recovers_tx_b_orphan_exactly_once_before_supervisor_ack() {
-        let (_dir, db_path, conn) = open_test_db_file();
-        let source_root = tempfile::TempDir::new().unwrap();
-        let pending_parent = tempfile::TempDir::new().unwrap();
-        let pending_root = pending_parent.path().join("restored");
-        let collection_id = insert_collection(&conn, "work", source_root.path());
-        write_restore_file(&pending_root, "notes/a.md", b"hello from restore");
-        let manifest_json = manifest_json_for_directory(&pending_root);
-        conn.execute(
-            "CREATE TABLE startup_finalize_audit (
-                 collection_id INTEGER NOT NULL,
-                 cleared_at TEXT NOT NULL DEFAULT (strftime('%Y-%m-%dT%H:%M:%SZ', 'now'))
-             )",
-            [],
-        )
-        .unwrap();
-        conn.execute(
-            "CREATE TRIGGER startup_finalize_exactly_once
-             AFTER UPDATE ON collections
-             WHEN OLD.pending_root_path IS NOT NULL AND NEW.pending_root_path IS NULL
-             BEGIN
-                 INSERT INTO startup_finalize_audit (collection_id) VALUES (NEW.id);
-             END",
-            [],
-        )
-        .unwrap();
-        conn.execute(
-            "INSERT INTO serve_sessions (session_id, pid, host, heartbeat_at)
-             VALUES ('stale-owner', 1, 'host', datetime('now', '-16 seconds'))",
-            [],
-        )
-        .unwrap();
-        conn.execute(
-            "INSERT INTO serve_sessions (session_id, pid, host)
-             VALUES ('foreign-live', 2, 'host')",
-            [],
-        )
-        .unwrap();
-        conn.execute(
-            "INSERT INTO collection_owners (collection_id, session_id)
-             VALUES (?1, 'stale-owner')",
-            [collection_id],
-        )
-        .unwrap();
-        conn.execute(
-            "UPDATE collections
-             SET state = 'restoring',
-                 pending_root_path = ?2,
-                 pending_restore_manifest = ?3,
-                 restore_command_id = 'restore-1',
-                 pending_command_heartbeat_at = datetime('now', '-16 seconds')
-             WHERE id = ?1",
-            params![
-                collection_id,
-                pending_root.display().to_string(),
-                manifest_json
-            ],
-        )
-        .unwrap();
-        drop(conn);
-
-        let runtime = start_serve_runtime(db_path.clone()).unwrap();
-
-        type StartupRecoveryStealsLeaseRow = (
-            String,
-            String,
-            i64,
-            Option<String>,
-            i64,
-            i64,
-            i64,
-            i64,
-            Option<String>,
-        );
-
-        let row: StartupRecoveryStealsLeaseRow = wait_for_collection_update(
-            &db_path,
-            collection_id,
-            Duration::from_secs(5),
-            |verify, collection_id| {
-                verify
-                        .query_row(
-                            "SELECT state,
-                                    root_path,
-                                    needs_full_sync,
-                                    pending_root_path,
-                                    (SELECT COUNT(*) FROM serve_sessions WHERE session_id = 'stale-owner'),
-                                    (SELECT COUNT(*) FROM serve_sessions WHERE session_id = 'foreign-live'),
-                                    (SELECT COUNT(*) FROM collection_owners WHERE collection_id = ?1 AND session_id = ?2),
-                                    (SELECT COUNT(*) FROM startup_finalize_audit WHERE collection_id = ?1),
-                                    watcher_released_session_id
-                             FROM collections
-                             WHERE id = ?1",
-                            params![collection_id, runtime.session_id.as_str()],
-                            |row| {
-                                Ok((
-                                    row.get(0)?,
-                                    row.get(1)?,
-                                    row.get(2)?,
-                                    row.get(3)?,
-                                    row.get(4)?,
-                                    row.get(5)?,
-                                    row.get(6)?,
-                                    row.get(7)?,
-                                    row.get(8)?,
-                                ))
-                            },
-                        )
-                        .ok()
-                        .and_then(|row| if row.0 == "active" { Some(row) } else { None })
-            },
-        );
-        assert_eq!(row.0, "active");
-        assert_eq!(row.1, pending_root.display().to_string());
-        assert_eq!(row.2, 0);
-        assert!(row.3.is_none());
-        assert_eq!(row.4, 0);
-        assert_eq!(row.5, 1);
-        assert_eq!(row.6, 1);
-        assert_eq!(row.7, 1);
-        assert!(row.8.is_none());
-
-        drop(runtime);
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn start_serve_runtime_defers_fresh_restore_heartbeat_and_leaves_collection_blocked() {
-        let (_dir, db_path, conn) = open_test_db_file();
-        let source_root = tempfile::TempDir::new().unwrap();
-        let source_root_canonical = fs::canonicalize(source_root.path()).unwrap();
-        let pending_parent = tempfile::TempDir::new().unwrap();
-        let pending_root = fs::canonicalize(pending_parent.path())
-            .unwrap()
-            .join("restored");
-        let collection_id = insert_collection(&conn, "work", &source_root_canonical);
-        write_restore_file(&pending_root, "notes/a.md", b"hello from restore");
-        let manifest_json = manifest_json_for_directory(&pending_root);
-        conn.execute(
-            "INSERT INTO serve_sessions (session_id, pid, host, heartbeat_at)
-             VALUES ('stale-owner', 1, 'host', datetime('now', '-16 seconds'))",
-            [],
-        )
-        .unwrap();
-        conn.execute(
-            "INSERT INTO collection_owners (collection_id, session_id)
-             VALUES (?1, 'stale-owner')",
-            [collection_id],
-        )
-        .unwrap();
-        conn.execute(
-            "UPDATE collections
-             SET state = 'restoring',
-                 pending_root_path = ?2,
-                 pending_restore_manifest = ?3,
-                 restore_command_id = 'restore-1',
-                 pending_command_heartbeat_at = strftime('%Y-%m-%dT%H:%M:%SZ', 'now')
-             WHERE id = ?1",
-            params![
-                collection_id,
-                pending_root.display().to_string(),
-                manifest_json
-            ],
-        )
-        .unwrap();
-        drop(conn);
-
-        let runtime = start_serve_runtime(db_path.clone()).unwrap();
-        thread::sleep(Duration::from_millis(500));
-
-        type RecoveryRow = (
-            String,
-            String,
-            i64,
-            Option<String>,
-            Option<String>,
-            i64,
-            Option<String>,
-            Option<i64>,
-            Option<String>,
-        );
-
-        let verify = Connection::open(&db_path).unwrap();
-        let row: RecoveryRow = verify
-            .query_row(
-                "SELECT state,
-                        root_path,
-                        needs_full_sync,
-                        pending_root_path,
-                        restore_command_id,
-                        (SELECT COUNT(*) FROM collection_owners WHERE collection_id = ?1 AND session_id = ?2),
-                        watcher_released_session_id,
-                        watcher_released_generation,
-                        watcher_released_at
-                 FROM collections
-                 WHERE id = ?1",
-                params![collection_id, runtime.session_id.as_str()],
-                |row| {
-                    Ok((
-                        row.get(0)?,
-                        row.get(1)?,
-                        row.get(2)?,
-                        row.get(3)?,
-                        row.get(4)?,
-                        row.get(5)?,
-                        row.get(6)?,
-                        row.get(7)?,
-                        row.get(8)?,
-                    ))
-                },
-            )
-            .unwrap();
-        assert_eq!(row.0, "restoring");
-        assert_eq!(row.1, source_root_canonical.display().to_string());
-        assert_eq!(row.2, 0);
-        assert_eq!(row.3.as_deref(), Some(pending_root.to_str().unwrap()));
-        assert_eq!(row.4.as_deref(), Some("restore-1"));
-        assert_eq!(row.5, 1);
-        assert!(
-            row.6.is_none() && row.7.is_none() && row.8.is_none(),
-            "fresh serve must not impersonate the originator by writing the watcher ack triple"
-        );
-
-        drop(runtime);
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn start_serve_runtime_bootstraps_recovery_directories_for_existing_collections() {
-        #[cfg(target_os = "linux")]
-        let _env_lock = env_mutation_lock().lock().unwrap();
-        #[cfg(target_os = "linux")]
-        let _runtime_root = secure_runtime_root();
-        #[cfg(target_os = "linux")]
-        let _xdg = EnvVarGuard::set("XDG_RUNTIME_DIR", _runtime_root.path().to_str().unwrap());
-        let (dir, db_path, conn) = open_test_db_file();
-        let root_a = tempfile::TempDir::new().unwrap();
-        let root_b = tempfile::TempDir::new().unwrap();
-        let collection_a = insert_collection(&conn, "work", root_a.path());
-        let collection_b = insert_collection(&conn, "notes", root_b.path());
-        drop(conn);
-
-        let runtime = start_serve_runtime(db_path).unwrap();
-
-        let recovery_root = dir.path().join("recovery");
-        assert!(collection_recovery_dir(&recovery_root, collection_a).is_dir());
-        assert!(collection_recovery_dir(&recovery_root, collection_b).is_dir());
-
-        drop(runtime);
     }
 
     #[cfg(unix)]
@@ -11620,334 +8491,6 @@ mod tests {
     }
 
     #[test]
-    fn write_supervisor_ack_rejects_foreign_stale_and_replayed_acknowledgements() {
-        let conn = open_test_db();
-        let temp = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", temp.path());
-        conn.execute(
-            "INSERT INTO serve_sessions (session_id, pid, host) VALUES ('serve-1', 1, 'host')",
-            [],
-        )
-        .unwrap();
-        conn.execute(
-            "INSERT INTO collection_owners (collection_id, session_id) VALUES (?1, 'serve-1')",
-            [collection_id],
-        )
-        .unwrap();
-        conn.execute(
-            "INSERT INTO serve_sessions (session_id, pid, host) VALUES ('serve-2', 2, 'host')",
-            [],
-        )
-        .unwrap();
-        conn.execute(
-            "UPDATE collections SET state = 'restoring', reload_generation = 2 WHERE id = ?1",
-            [collection_id],
-        )
-        .unwrap();
-
-        assert!(!write_supervisor_ack_if_needed(&conn, collection_id, "serve-2", 2).unwrap());
-        assert!(!write_supervisor_ack_if_needed(&conn, collection_id, "serve-1", 1).unwrap());
-        assert!(write_supervisor_ack_if_needed(&conn, collection_id, "serve-1", 2).unwrap());
-        assert!(!write_supervisor_ack_if_needed(&conn, collection_id, "serve-1", 2).unwrap());
-
-        let ack: (Option<String>, Option<i64>) = conn
-            .query_row(
-                "SELECT watcher_released_session_id, watcher_released_generation
-                 FROM collections WHERE id = ?1",
-                [collection_id],
-                |row| Ok((row.get(0)?, row.get(1)?)),
-            )
-            .unwrap();
-        assert_eq!(ack.0.as_deref(), Some("serve-1"));
-        assert_eq!(ack.1, Some(2));
-    }
-
-    #[test]
-    fn wait_for_exact_ack_reports_when_serve_ownership_changes_mid_handshake() {
-        let conn = open_test_db();
-        let temp = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", temp.path());
-        conn.execute(
-            "INSERT INTO serve_sessions (session_id, pid, host, heartbeat_at)
-             VALUES ('serve-1', 1, 'host', datetime('now'))",
-            [],
-        )
-        .unwrap();
-        conn.execute(
-            "INSERT INTO serve_sessions (session_id, pid, host, heartbeat_at)
-             VALUES ('serve-2', 2, 'host', datetime('now'))",
-            [],
-        )
-        .unwrap();
-        conn.execute(
-            "INSERT INTO collection_owners (collection_id, session_id) VALUES (?1, 'serve-2')",
-            [collection_id],
-        )
-        .unwrap();
-
-        match wait_for_exact_ack(&conn, collection_id, "serve-1", 2) {
-            Err(VaultSyncError::ServeOwnershipChanged {
-                collection_name,
-                expected_session_id,
-                actual_session_id,
-            }) => {
-                assert_eq!(collection_name, "work");
-                assert_eq!(expected_session_id, "serve-1");
-                assert_eq!(actual_session_id, "serve-2");
-            }
-            other => panic!("expected ServeOwnershipChangedError, got {other:?}"),
-        }
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn full_hash_audit_pass_rehashes_due_active_lease_collections_only() {
-        let conn = open_test_db();
-        let due_root = tempfile::TempDir::new().unwrap();
-        let skipped_root = tempfile::TempDir::new().unwrap();
-        let due_id = insert_collection(&conn, "due", due_root.path());
-        let skipped_id = insert_collection(&conn, "skipped", skipped_root.path());
-
-        let due_uuid = Uuid::now_v7().to_string();
-        let due_bytes = format!(
-            "---\nmemory_id: {due_uuid}\nslug: notes/due\ntitle: Due\ntype: concept\n---\nBody.\n"
-        );
-        fs::create_dir_all(due_root.path().join("notes")).unwrap();
-        fs::write(due_root.path().join("notes/due.md"), due_bytes.as_bytes()).unwrap();
-        insert_page_with_raw_import(
-            &conn,
-            due_id,
-            "notes/due",
-            &due_uuid,
-            "Body.",
-            due_bytes.as_bytes(),
-            "notes/due.md",
-        );
-        conn.execute(
-            "UPDATE file_state
-             SET last_full_hash_at = datetime('now', '-8 days')
-             WHERE collection_id = ?1",
-            [due_id],
-        )
-        .unwrap();
-
-        let skipped_uuid = Uuid::now_v7().to_string();
-        let skipped_bytes = format!(
-            "---\nmemory_id: {skipped_uuid}\nslug: notes/skipped\ntitle: Skipped\ntype: concept\n---\nBody.\n"
-        );
-        fs::create_dir_all(skipped_root.path().join("notes")).unwrap();
-        fs::write(
-            skipped_root.path().join("notes/skipped.md"),
-            skipped_bytes.as_bytes(),
-        )
-        .unwrap();
-        insert_page_with_raw_import(
-            &conn,
-            skipped_id,
-            "notes/skipped",
-            &skipped_uuid,
-            "Body.",
-            skipped_bytes.as_bytes(),
-            "notes/skipped.md",
-        );
-        conn.execute(
-            "UPDATE file_state
-             SET last_full_hash_at = datetime('now', '-1 days')
-             WHERE collection_id = ?1",
-            [skipped_id],
-        )
-        .unwrap();
-
-        conn.execute(
-            "INSERT INTO serve_sessions (session_id, pid, host, heartbeat_at)
-             VALUES ('serve-audit', 1, 'host', datetime('now'))",
-            [],
-        )
-        .unwrap();
-        conn.execute(
-            "INSERT INTO collection_owners (collection_id, session_id) VALUES (?1, 'serve-audit')",
-            [due_id],
-        )
-        .unwrap();
-        conn.execute(
-            "INSERT INTO collection_owners (collection_id, session_id) VALUES (?1, 'serve-audit')",
-            [skipped_id],
-        )
-        .unwrap();
-        conn.execute(
-            "UPDATE collections SET active_lease_session_id = 'serve-audit' WHERE id IN (?1, ?2)",
-            rusqlite::params![due_id, skipped_id],
-        )
-        .unwrap();
-
-        let audited = run_full_hash_audit_pass(&conn, "serve-audit").unwrap();
-
-        assert_eq!(audited.len(), 1);
-        assert_eq!(audited[0].0, due_id);
-        assert_eq!(audited[0].1, "due");
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn full_hash_audit_pass_limits_each_cycle_to_a_daily_subset() {
-        let _guard = env_mutation_lock().lock().unwrap();
-        let _env = EnvVarGuard::set("QUAID_FULL_HASH_AUDIT_DAYS", "3");
-        let conn = open_test_db();
-        let root = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", root.path());
-
-        fs::create_dir_all(root.path().join("notes")).unwrap();
-        for index in 0..7 {
-            let relative_path = format!("notes/{index:02}.md");
-            let slug = format!("notes/{index:02}");
-            let uuid = Uuid::now_v7().to_string();
-            let raw = format!(
-                "---\nmemory_id: {uuid}\nslug: {slug}\ntitle: Note {index}\ntype: concept\n---\nBody {index}.\n"
-            );
-            fs::write(root.path().join(&relative_path), raw.as_bytes()).unwrap();
-            insert_page_with_raw_import(
-                &conn,
-                collection_id,
-                &slug,
-                &uuid,
-                &format!("Body {index}."),
-                raw.as_bytes(),
-                &relative_path,
-            );
-        }
-        conn.execute(
-            "UPDATE file_state
-             SET last_full_hash_at = datetime('now', '-8 days')
-             WHERE collection_id = ?1",
-            [collection_id],
-        )
-        .unwrap();
-        conn.execute(
-            "INSERT INTO serve_sessions (session_id, pid, host, heartbeat_at)
-             VALUES ('serve-audit', 1, 'host', datetime('now'))",
-            [],
-        )
-        .unwrap();
-        conn.execute(
-            "INSERT INTO collection_owners (collection_id, session_id) VALUES (?1, 'serve-audit')",
-            [collection_id],
-        )
-        .unwrap();
-        conn.execute(
-            "UPDATE collections SET active_lease_session_id = 'serve-audit' WHERE id = ?1",
-            [collection_id],
-        )
-        .unwrap();
-
-        let audited = run_full_hash_audit_pass(&conn, "serve-audit").unwrap();
-
-        assert_eq!(audited.len(), 1);
-        assert_eq!(
-            audited[0].2.walked, 3,
-            "7 files over 3 days should hash only ceil(7/3) files per cycle"
-        );
-        assert_eq!(audited[0].2.unchanged, 3);
-        let fresh_paths = conn
-            .prepare(
-                "SELECT relative_path
-                 FROM file_state
-                 WHERE collection_id = ?1
-                   AND last_full_hash_at > datetime('now', '-1 minute')
-                 ORDER BY relative_path ASC",
-            )
-            .unwrap()
-            .query_map([collection_id], |row| row.get::<_, String>(0))
-            .unwrap()
-            .collect::<Result<Vec<_>, _>>()
-            .unwrap();
-        let stale_paths = conn
-            .prepare(
-                "SELECT relative_path
-                 FROM file_state
-                 WHERE collection_id = ?1
-                   AND last_full_hash_at < datetime('now', '-7 days')
-                 ORDER BY relative_path ASC",
-            )
-            .unwrap()
-            .query_map([collection_id], |row| row.get::<_, String>(0))
-            .unwrap()
-            .collect::<Result<Vec<_>, _>>()
-            .unwrap();
-        assert_eq!(
-            fresh_paths,
-            vec![
-                "notes/00.md".to_owned(),
-                "notes/01.md".to_owned(),
-                "notes/02.md".to_owned()
-            ],
-            "scheduled audit must update only the first budgeted subset this cycle"
-        );
-        assert_eq!(stale_paths.len(), 4);
-
-        let audited_again = run_full_hash_audit_pass(&conn, "serve-audit").unwrap();
-
-        assert_eq!(audited_again.len(), 1);
-        assert_eq!(
-            audited_again[0].2.walked, 3,
-            "the next serve-loop cycle must stay bounded to the same daily subset size"
-        );
-        let remaining_stale = conn
-            .prepare(
-                "SELECT relative_path
-                 FROM file_state
-                 WHERE collection_id = ?1
-                   AND last_full_hash_at < datetime('now', '-7 days')
-                 ORDER BY relative_path ASC",
-            )
-            .unwrap()
-            .query_map([collection_id], |row| row.get::<_, String>(0))
-            .unwrap()
-            .collect::<Result<Vec<_>, _>>()
-            .unwrap();
-        assert_eq!(
-            remaining_stale,
-            vec!["notes/06.md".to_owned()],
-            "scheduled audit must advance to the next oldest subset instead of re-running a whole-vault pass"
-        );
-    }
-
-    #[test]
-    fn run_full_hash_audit_pass_batches_due_rows_instead_of_inline_full_vault_reconcile() {
-        let source = fs::read_to_string(
-            Path::new(env!("CARGO_MANIFEST_DIR"))
-                .join("src")
-                .join("core")
-                .join("vault_sync.rs"),
-        )
-        .unwrap();
-        let fn_start = source
-            .find("pub fn run_full_hash_audit_pass(")
-            .expect("run_full_hash_audit_pass fn present");
-        let fn_end = source[fn_start..]
-            .find("pub fn audit_collection(")
-            .map(|offset| fn_start + offset)
-            .expect("audit_collection fn follows run_full_hash_audit_pass");
-        let fn_body = &source[fn_start..fn_end];
-
-        assert!(
-            fn_body.contains("scheduled_full_hash_audit_budget(total_files.max(0) as usize)"),
-            "scheduled audit must compute a bounded per-cycle budget before hashing overdue rows"
-        );
-        assert!(
-            fn_body.contains("LIMIT ?3"),
-            "scheduled audit must select only the budgeted overdue rows per serve-loop cycle"
-        );
-        assert!(
-            fn_body.contains("scheduled_full_hash_audit_authorized("),
-            "scheduled audit must hash only the selected overdue subset"
-        );
-        assert!(
-            !fn_body.contains("full_hash_reconcile_authorized("),
-            "run_full_hash_audit_pass must not collapse back into a whole-vault inline reconcile"
-        );
-    }
-
-    #[test]
     fn full_hash_audit_helpers_respect_env_override_and_floor_at_zero() {
         let _guard = env_mutation_lock().lock().unwrap();
 
@@ -12078,59 +8621,6 @@ mod tests {
     }
 
     #[test]
-    fn live_collection_owner_ignores_stale_heartbeat_rows_older_than_fifteen_seconds() {
-        let conn = open_test_db();
-        let temp = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", temp.path());
-        conn.execute(
-            "INSERT INTO serve_sessions (session_id, pid, host, heartbeat_at)
-             VALUES ('serve-stale', 77, 'host', datetime('now', '-16 seconds'))",
-            [],
-        )
-        .unwrap();
-        conn.execute(
-            "INSERT INTO collection_owners (collection_id, session_id) VALUES (?1, 'serve-stale')",
-            [collection_id],
-        )
-        .unwrap();
-
-        assert!(live_collection_owner(&conn, collection_id)
-            .unwrap()
-            .is_none());
-    }
-
-    #[test]
-    fn start_serve_runtime_refreshes_session_heartbeat_on_five_second_interval() {
-        let (_dir, db_path, conn) = open_test_db_file();
-        drop(conn);
-
-        let runtime = start_serve_runtime(db_path.clone()).unwrap();
-        let conn = Connection::open(&db_path).unwrap();
-        let first_heartbeat: String = conn
-            .query_row(
-                "SELECT heartbeat_at FROM serve_sessions WHERE session_id = ?1",
-                [runtime.session_id.as_str()],
-                |row| row.get(0),
-            )
-            .unwrap();
-        drop(conn);
-
-        thread::sleep(Duration::from_millis(6200));
-
-        let conn = Connection::open(&db_path).unwrap();
-        let second_heartbeat: String = conn
-            .query_row(
-                "SELECT heartbeat_at FROM serve_sessions WHERE session_id = ?1",
-                [runtime.session_id.as_str()],
-                |row| row.get(0),
-            )
-            .unwrap();
-        assert_ne!(first_heartbeat, second_heartbeat);
-
-        drop(runtime);
-    }
-
-    #[test]
     fn short_lived_owner_lease_releases_on_panic_unwind() {
         let (_dir, _db_path, conn) = open_test_db_file();
         let temp = tempfile::TempDir::new().unwrap();
@@ -12256,79 +8746,6 @@ mod tests {
         );
     }
 
-    #[test]
-    fn ensure_no_live_serve_owner_for_root_path_reports_same_root_alias_owner() {
-        let conn = open_test_db();
-        let temp = tempfile::TempDir::new().unwrap();
-        let temp_canonical = fs::canonicalize(temp.path()).unwrap();
-        let collection_id = insert_collection(&conn, "work", &temp_canonical);
-        insert_collection(&conn, "alias", &temp_canonical);
-        conn.execute(
-            "INSERT INTO serve_sessions (session_id, pid, host, heartbeat_at)
-             VALUES ('serve-live', 77, 'batch3-host', datetime('now'))",
-            [],
-        )
-        .unwrap();
-        conn.execute(
-            "INSERT INTO collection_owners (collection_id, session_id) VALUES (?1, 'serve-live')",
-            [collection_id],
-        )
-        .unwrap();
-
-        let error =
-            ensure_no_live_serve_owner_for_root_path(&conn, &temp_canonical.display().to_string())
-                .unwrap_err();
-        let text = error.to_string();
-        assert!(text.contains("ServeOwnsCollectionError"));
-        assert!(text.contains("collection=work"));
-        assert!(text.contains("owner_pid=77"));
-        assert!(text.contains("owner_host=batch3-host"));
-    }
-
-    #[test]
-    fn ensure_no_live_serve_owner_for_root_path_allows_stale_same_root_owner_residue() {
-        let conn = open_test_db();
-        let temp = tempfile::TempDir::new().unwrap();
-        insert_collection(&conn, "work", temp.path());
-        let alias_id = insert_collection(&conn, "alias", temp.path());
-        conn.execute(
-            "INSERT INTO serve_sessions (session_id, pid, host, heartbeat_at)
-             VALUES ('serve-stale', 77, 'batch3-host', datetime('now', '-120 seconds'))",
-            [],
-        )
-        .unwrap();
-        conn.execute(
-            "INSERT INTO collection_owners (collection_id, session_id) VALUES (?1, 'serve-stale')",
-            [alias_id],
-        )
-        .unwrap();
-
-        ensure_no_live_serve_owner_for_root_path(&conn, &temp.path().display().to_string())
-            .unwrap();
-    }
-
-    #[test]
-    fn ensure_no_live_serve_owner_for_root_path_ignores_cli_session() {
-        let conn = open_test_db();
-        let temp = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", temp.path());
-        conn.execute(
-            "INSERT INTO serve_sessions (session_id, pid, host, heartbeat_at, session_type)
-             VALUES ('cli-lease', 99, 'cli-host', datetime('now'), 'cli')",
-            [],
-        )
-        .unwrap();
-        conn.execute(
-            "INSERT INTO collection_owners (collection_id, session_id) VALUES (?1, 'cli-lease')",
-            [collection_id],
-        )
-        .unwrap();
-
-        // A live CLI-type lease must not trigger a ServeOwnsCollectionError.
-        ensure_no_live_serve_owner_for_root_path(&conn, &temp.path().display().to_string())
-            .unwrap();
-    }
-
     /// Regression: during the shutdown race window the serve session's `ipc_path`
     /// column may be cleared (by `cleanup_published_ipc_socket`) before the session
     /// row is unregistered.  `live_serve_endpoint_for_root_path` must treat a NULL
@@ -12360,67 +8777,6 @@ mod tests {
         assert!(
             matches!(result, Ok(None)),
             "expected Ok(None) for NULL ipc_path during shutdown race, got: {result:?}"
-        );
-    }
-
-    /// Regression: `accept_ipc_clients` must offload each accepted connection to a
-    /// dedicated thread via `thread::spawn` so that blocking IPC I/O (up to 5 s per
-    /// read/write timeout) cannot stall the main serve loop and trigger a false-dead
-    /// live-owner verdict.  We verify the structural invariant via source inspection.
-    #[test]
-    fn accept_ipc_clients_offloads_each_client_to_its_own_thread() {
-        let source = fs::read_to_string(
-            Path::new(env!("CARGO_MANIFEST_DIR"))
-                .join("src")
-                .join("core")
-                .join("vault_sync.rs"),
-        )
-        .unwrap();
-        let fn_start = source
-            .find("fn accept_ipc_clients(")
-            .expect("accept_ipc_clients fn present");
-        // Boundary: the next fn after accept_ipc_clients is handle_ipc_client.
-        let fn_end = source[fn_start..]
-            .find("fn handle_ipc_client(")
-            .map(|offset| fn_start + offset)
-            .expect("handle_ipc_client fn follows accept_ipc_clients");
-        let fn_body = &source[fn_start..fn_end];
-
-        // The function must not call handle_ipc_client directly on the main loop thread.
-        assert!(
-            !fn_body.contains("handle_ipc_client(stream,"),
-            "accept_ipc_clients must not call handle_ipc_client inline on the main thread"
-        );
-        // The function must use thread::spawn to offload.
-        assert!(
-            fn_body.contains("thread::spawn("),
-            "accept_ipc_clients must offload each client via thread::spawn"
-        );
-        // The spawn closure must contain the handle_ipc_client call.
-        let spawn_idx = fn_body.find("thread::spawn(").unwrap();
-        let spawn_region = &fn_body[spawn_idx..];
-        assert!(
-            spawn_region.contains("handle_ipc_client("),
-            "handle_ipc_client must be called inside the thread::spawn closure"
-        );
-        // The function must enforce the in-flight cap before spawning.
-        assert!(
-            fn_body.contains("IPC_HANDLER_LIMIT"),
-            "accept_ipc_clients must reference IPC_HANDLER_LIMIT"
-        );
-        assert!(
-            fn_body.contains("fetch_add("),
-            "accept_ipc_clients must use fetch_add for the in-flight counter"
-        );
-        // Rollback path: fetch_sub must appear for the saturation branch.
-        assert!(
-            fn_body.contains("fetch_sub("),
-            "accept_ipc_clients must roll back via fetch_sub when saturated"
-        );
-        // Guard must be created before the spawn closure.
-        assert!(
-            fn_body.contains("IpcHandlerGuard("),
-            "accept_ipc_clients must construct an IpcHandlerGuard before spawning"
         );
     }
 
@@ -12459,143 +8815,4 @@ mod tests {
         );
     }
 
-    #[cfg(unix)]
-    #[test]
-    fn plain_sync_reconciles_active_root_and_clears_needs_full_sync() {
-        let (_dir, _db_path, conn) = open_test_db_file();
-        let root = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", root.path());
-        conn.execute(
-            "UPDATE collections SET needs_full_sync = 1 WHERE id = ?1",
-            [collection_id],
-        )
-        .unwrap();
-        fs::write(
-            root.path().join("note.md"),
-            "---\nslug: notes/note\ntitle: Note\ntype: concept\n---\nA body long enough to reconcile through the active-root path.\n",
-        )
-        .unwrap();
-
-        let stats = sync_collection(&conn, "work").unwrap();
-        let row: (String, i64, Option<String>, Option<String>) = conn
-            .query_row(
-                "SELECT state, needs_full_sync, last_sync_at, active_lease_session_id
-                 FROM collections WHERE id = ?1",
-                [collection_id],
-                |row| Ok((row.get(0)?, row.get(1)?, row.get(2)?, row.get(3)?)),
-            )
-            .unwrap();
-        let page_count: i64 = conn
-            .query_row(
-                "SELECT COUNT(*) FROM pages WHERE collection_id = ?1",
-                [collection_id],
-                |row| row.get(0),
-            )
-            .unwrap();
-
-        assert_eq!(stats.walked, 1);
-        assert_eq!(row.0, "active");
-        assert_eq!(row.1, 0);
-        assert!(row.2.is_some());
-        assert!(row.3.is_none());
-        assert_eq!(page_count, 1);
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn plain_sync_turns_duplicate_uuid_into_terminal_reconcile_halt() {
-        let (_dir, _db_path, conn) = open_test_db_file();
-        let root = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", root.path());
-        let uuid = "01969f11-9448-7d79-8d3f-c68f54768888";
-        let note_a = format!(
-            "---\nmemory_id: {uuid}\nslug: notes/a\ntitle: A\ntype: concept\n---\nThis body is comfortably above the minimum size for rename inference.\n"
-        );
-        let note_b = format!(
-            "---\nmemory_id: {uuid}\nslug: notes/b\ntitle: B\ntype: concept\n---\nThis second body is also comfortably above the minimum size for rename inference.\n"
-        );
-        fs::write(root.path().join("a.md"), note_a).unwrap();
-        fs::write(root.path().join("b.md"), note_b).unwrap();
-
-        let error = sync_collection(&conn, "work").unwrap_err().to_string();
-        let row: (Option<String>, Option<String>, i64) = conn
-            .query_row(
-                "SELECT reconcile_halted_at, reconcile_halt_reason,
-                        (SELECT COUNT(*) FROM pages WHERE collection_id = ?1)
-                 FROM collections WHERE id = ?1",
-                [collection_id],
-                |row| Ok((row.get(0)?, row.get(1)?, row.get(2)?)),
-            )
-            .unwrap();
-
-        assert!(error.contains("ReconcileHaltedError"));
-        assert!(error.contains("DuplicateUuidError"));
-        assert!(row.0.is_some());
-        assert_eq!(row.1.as_deref(), Some("duplicate_uuid"));
-        assert_eq!(row.2, 0);
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn plain_sync_turns_trivial_hash_ambiguity_into_terminal_reconcile_halt() {
-        let (_dir, _db_path, conn) = open_test_db_file();
-        let root = tempfile::TempDir::new().unwrap();
-        let collection_id = insert_collection(&conn, "work", root.path());
-        let content = concat!(
-            "---\n",
-            "slug: notes/template\n",
-            "title: Template Note\n",
-            "type: concept\n",
-            "meta: aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa\n",
-            "---\n",
-            "Hi\n",
-        );
-        fs::write(root.path().join("template.md"), content).unwrap();
-        let stat = crate::core::file_state::stat_file(&root.path().join("template.md")).unwrap();
-        let sha256 = crate::core::file_state::hash_file(&root.path().join("template.md")).unwrap();
-        conn.execute(
-            "INSERT INTO pages (collection_id, slug, uuid, type, title, compiled_truth, timeline)
-             VALUES (?1, 'notes/template', ?2, 'concept', 'Template', 'Hi', '')",
-            params![collection_id, "01969f11-9448-7d79-8d3f-c68f54767777"],
-        )
-        .unwrap();
-        let page_id = conn.last_insert_rowid();
-        conn.execute(
-            "INSERT INTO file_state
-                 (collection_id, relative_path, page_id, mtime_ns, ctime_ns, size_bytes, inode, sha256)
-             VALUES (?1, 'old-template.md', ?2, ?3, ?4, ?5, ?6, ?7)",
-            params![
-                collection_id,
-                page_id,
-                stat.mtime_ns,
-                stat.ctime_ns,
-                stat.size_bytes,
-                stat.inode,
-                sha256
-            ],
-        )
-        .unwrap();
-        crate::core::raw_imports::rotate_active_raw_import(
-            &conn,
-            page_id,
-            "old-template.md",
-            content.as_bytes(),
-        )
-        .unwrap();
-
-        let error = sync_collection(&conn, "work").unwrap_err().to_string();
-        let row: (Option<String>, Option<String>) = conn
-            .query_row(
-                "SELECT reconcile_halted_at, reconcile_halt_reason
-                 FROM collections WHERE id = ?1",
-                [collection_id],
-                |row| Ok((row.get(0)?, row.get(1)?)),
-            )
-            .unwrap();
-
-        assert!(error.contains("ReconcileHaltedError"));
-        assert!(error.contains("UnresolvableTrivialContentError"));
-        assert!(row.0.is_some());
-        assert_eq!(row.1.as_deref(), Some("unresolvable_trivial_content"));
-    }
 }
